% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddCardModule.R
\name{add_card_button_ui}
\alias{add_card_button_ui}
\title{Add Card Button User Interface}
\usage{
add_card_button_ui(id)
}
\arguments{
\item{id}{\code{character}}
}
\value{
\code{shiny::tagList}
}
\description{
button for adding views/cards to the Report. Part of the simple Reporter user interface.

For more details see the vignette: \code{vignette("simpleReporter", "teal.reporter")}.
}
