% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_report-class.R
\name{append_document}
\alias{append_document}
\title{Append elements to a teal_report document}
\usage{
append_document(x, values, after = length(document(x)))
}
\arguments{
\item{x}{A teal_report object}

\item{values}{(\code{teal_document}) object(s) to append}

\item{after}{Integer, the position after which the elements are to be appended.
If negative or zero, the values are prepended to the document.
If missing, the values are appended at the end.}
}
\value{
A teal_report object with updated document slot
}
\description{
This function appends elements to the document slot of a teal_report object.
It follows the same pattern as base R's append() function.
}
\examples{
data <- teal_report()
data <- append_document(data, teal_document("Some text"))
data <- append_document(data, list(
  teal_document("More text"),
  teal_document("Even more text")
))
data <- append_document(data, teal_document("First text"), after = 0)
document(data)

}
