% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchiverModule.R
\name{archiver_save_srv}
\alias{archiver_save_srv}
\title{Archiver Server}
\usage{
archiver_save_srv(id, reporter)
}
\arguments{
\item{id}{\code{character(1)} this \code{shiny} module's id.}

\item{reporter}{\code{\link{Reporter}} instance.}
}
\value{
\code{shiny::moduleServer}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
server for saving and loading to the Report.

For more details see the vignette: \code{vignette("simpleReporter", "teal.reporter")}.
}
