% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_report-class.R
\name{as_teal_document}
\alias{as_teal_document}
\title{Create or coerce to a teal_document}
\usage{
as_teal_document(x)
}
\arguments{
\item{x}{Object to convert to teal_document}
}
\value{
A teal_document object
}
\description{
This function ensures that input is converted to a teal_document object.
It accepts various input types and converts them appropriately.
}
