% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportDocument.R
\name{block_to_markdown}
\alias{block_to_markdown}
\alias{block_to_markdown.default}
\alias{block_to_markdown.ggplot}
\title{#' @export
#' @rdname code_output
link_output <- function(object, output) {
attr(object, "output") <- output
object
}}
\usage{
block_to_markdown(x)

\method{block_to_markdown}{default}(x)

\method{block_to_markdown}{ggplot}(x, width = 5, height = 4, dpi = 100)
}
\description{
#' @export
#' @rdname code_output
link_output <- function(object, output) {
attr(object, "output") <- output
object
}
}
