% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{card}
\alias{card}
\alias{card<-}
\alias{as.card}
\alias{c.card}
\alias{[.card}
\alias{edit_card}
\title{\code{card}: An \code{S3} class for managing \code{teal} reports}
\usage{
card(x, ...)

card(x) <- value

as.card(x)

\method{c}{card}(...)

\method{[}{card}(x, i)

edit_card(x, modify = NULL, append = NULL, after = length(x))
}
\arguments{
\item{x}{\code{card}}

\item{...}{Additional elements to include when creating a new \code{card}}

\item{i}{index specifying elements to extract or replace}

\item{modify}{An integer vector specifying element indices to extract and reorder.
If \code{NULL}, no modification is applied.}

\item{append}{An object to be added to the \code{card} using \code{append()}.
The \code{after} parameter determines the insertion position.}

\item{after}{a subscript, after which the values are to be appended.}
}
\value{
An \code{S3} \code{list} of class \code{card}.

A card object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The \code{card} \code{S3} class provides functionality to store, manage, edit, and adjust report contents.
It enables users to create, manipulate, and serialize report-related data efficiently.

The \code{card()} function serves two purposes:
\enumerate{
\item When called with a \code{teal_report} object, it acts as a getter and returns the card slot
\item When called with other arguments, it creates a new \code{card} object from those arguments
}

This function ensures that input is converted to a card object.
It accepts various input types and converts them appropriately.
}
\details{
The \code{card} class supports \code{c()} and \code{x[i]} methods for combining and subsetting elements.
However, these methods only function correctly when the first element is a \code{card}.
To prepend, reorder, or modify a \code{card}, use the \code{edit_card()} function.
}
\examples{
# Create a new empty card
report <- card()
class(report) # Check the class of the object

# Create a card with content
report <- card("## Headline", "Some text", summary(iris))

# Extract card from a teal_report
tr <- teal_report(card = card("## Title"))
doc <- card(tr)

# Add elements to the report
report <- c(report, list("## Table"), list(summary(mtcars)))

# Subset the report to keep only the first two elements
report <- report[1:2]

# Append new elements after the first element
report <- append(report, c(list("## Table 2"), list(summary(mtcars))), after = 1)

# Verify that the object remains a card
class(report)

#### edit_card examples ###
report <- card(1, 2, "c")

# Modify and append to the report
new_report <- edit_card(report, modify = c(3, 1), append = "d")
new_report
class(new_report)

}
