% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportDocument.R
\name{code_chunk}
\alias{code_chunk}
\alias{code_output}
\title{Generate an R Markdown code chunk}
\usage{
code_chunk(code, ...)

code_output(code)
}
\arguments{
\item{code}{A character string containing the R code to be wrapped in the chunk.}

\item{...}{Additional named parameters to be included inside \code{{r}}.}
}
\value{
A formatted character string representing an R Markdown code chunk.
}
\description{
This function takes a character string as input and formats it as an R Markdown code chunk.
Additional named parameters passed via \code{...} will be included inside \code{{r}}.
}
\examples{
code_chunk("x <- 1:10", echo = TRUE, message = FALSE)
}
