% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContentBlock.R
\name{ContentBlock}
\alias{ContentBlock}
\title{\code{ContentBlock}}
\description{
\code{ContentBlock}

\code{ContentBlock}
}
\examples{

## ------------------------------------------------
## Method `ContentBlock$new`
## ------------------------------------------------

block <- teal.reporter:::ContentBlock$new()


## ------------------------------------------------
## Method `ContentBlock$set_content`
## ------------------------------------------------

block <- teal.reporter:::ContentBlock$new()
block$set_content("Base64 encoded picture")


## ------------------------------------------------
## Method `ContentBlock$get_content`
## ------------------------------------------------

block <- teal.reporter:::ContentBlock$new()
block$get_content()

}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ContentBlock$new()}}
\item \href{#method-set_content}{\code{ContentBlock$set_content()}}
\item \href{#method-get_content}{\code{ContentBlock$get_content()}}
\item \href{#method-clone}{\code{ContentBlock$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Returns a \code{ContentBlock} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentBlock$new()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Returns a \code{ContentBlock} object with no content and the default style.
}

\subsection{Returns}{
\code{ContentBlock}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{block <- teal.reporter:::ContentBlock$new()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_content"></a>}}
\if{latex}{\out{\hypertarget{method-set_content}{}}}
\subsection{Method \code{set_content()}}{
Sets content of this \code{ContentBlock}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentBlock$set_content(content)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{content}}{(\code{character(0)} or \code{character(1)}) a string literal or a file path assigned to this \code{ContentBlock}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisibly self
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{block <- teal.reporter:::ContentBlock$new()
block$set_content("Base64 encoded picture")

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_content"></a>}}
\if{latex}{\out{\hypertarget{method-get_content}{}}}
\subsection{Method \code{get_content()}}{
Returns the absolute path to content of this \code{ContentBlock}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentBlock$get_content()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} content of this \code{ContentBlock}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{block <- teal.reporter:::ContentBlock$new()
block$get_content()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContentBlock$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
