% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_utils.R
\name{conv_str_logi}
\alias{conv_str_logi}
\title{Convert a character of a yaml boolean to a logical value}
\usage{
conv_str_logi(
  input,
  pos_logi = c("TRUE", "true", "True", "yes", "y", "Y", "on"),
  neg_logi = c("FALSE", "false", "False", "no", "n", "N", "off"),
  silent = FALSE
)
}
\arguments{
\item{input}{\code{character}}

\item{pos_logi}{\code{character} vector}

\item{neg_logi}{\code{character} vector}

\item{silent}{\code{logical}}
}
\value{
\code{input} argument or the appropriate \code{logical} value.
}
\description{
convert a character of a yaml boolean to a logical value.
}
\examples{
teal.reporter:::conv_str_logi("TRUE")
teal.reporter:::conv_str_logi("True")

teal.reporter:::conv_str_logi("off")
teal.reporter:::conv_str_logi("n")

teal.reporter:::conv_str_logi("sth")

}
\keyword{internal}
