% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_report-report.R
\name{document}
\alias{document}
\alias{document<-}
\title{Get or set the document of a \code{teal_report} object}
\usage{
document(x)

document(x) <- value
}
\arguments{
\item{x}{(\code{teal_report})}

\item{value}{(\code{teal_document})}
}
\value{
The document of the \code{teal_report} object.
}
\description{
Get or set the document of a \code{teal_report} object
}
