% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.action_button_busy}
\alias{.action_button_busy}
\title{Teal action button that is disabled while busy}
\usage{
.action_button_busy(
  id,
  label,
  icon = NULL,
  type = "primary",
  outline = FALSE,
  additional_class = NULL
)
}
\arguments{
\item{id}{(\code{character(1)}) the id of the button.}

\item{label}{(\code{character(1)}) the label of the button.}

\item{icon}{(\code{character(1)} or \code{NULL}) the name of the Bootstrap icon to be
displayed on the button.}

\item{type}{One of the Bootstrap theme colors (\code{"primary"}, \code{"default"},
\code{"secondary"}, \code{"success"}, \code{"danger"}, \code{"warning"}, \code{"info"}, \code{"light"},
\code{"dark"}), or \code{NULL} to leave off the Bootstrap-specific button CSS classes
altogether.}

\item{additional_class}{(\code{character(1)} or \code{NULL}) additional CSS class to be
added to the button.}
}
\value{
A \code{shiny} action button that is disabled while busy.
}
\description{
Teal action button that is disabled while busy
}
\keyword{internal}
