% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_card.R
\name{.build_card_from_code}
\alias{.build_card_from_code}
\title{Builds \code{teal_card} from code and outputs in \code{qenv} object}
\usage{
.build_card_from_code(data)
}
\arguments{
\item{data}{(\code{qenv}) object.}
}
\value{
A \code{teal_card} built from the code and outputs in a \code{qenv}
object.
}
\description{
Builds a \code{teal_card} from the code and outputs of a \code{teal_data}
object, preserving the order of code execution and output display.
}
\keyword{internal}
