% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_card.R
\name{.code_to_card}
\alias{.code_to_card}
\title{Builds \code{teal_card} from code and outputs in \code{qenv} object}
\usage{
.code_to_card(x, code_block_opts = list())
}
\arguments{
\item{x}{(\code{list}) object from \code{qenv@code}.}

\item{code_block_opts}{(\code{list}) Additional options for the R code chunk in R Markdown.}
}
\value{
A \code{teal_card} built from the code and outputs in a \code{qenv} object.
}
\description{
Builds a \code{teal_card} from the code and outputs of a \code{teal_data}
object, preserving the order of code execution and output display.
}
\keyword{internal}
