% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_card.R
\name{.determine_default_dimensions}
\alias{.determine_default_dimensions}
\title{Determine default dimensions for report figures}
\usage{
.determine_default_dimensions(x, convert_to_inches = FALSE, dpi = 96)
}
\arguments{
\item{x}{An object, typically a \code{recordedplot} or \code{ggplot}, that has an
optional attributes \code{dev.width} and \code{dev.height} that override the default
dims set as options \code{teal.reporter.dev.fig.width} and
\code{teal.reporter.dev.fig.height}.}
}
\value{
List with \code{width} and \code{height} elements.
}
\description{
Determine default dimensions for report figures
}
\keyword{internal}
