% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadModule.R
\name{download_report_button_ui}
\alias{download_report_button_ui}
\title{Download Button Reporter User Interface}
\usage{
download_report_button_ui(id)
}
\arguments{
\item{id}{\code{character}}
}
\value{
\code{shiny::tagList}
}
\description{
button for downloading the Report. Part of the simple Reporter user interface.

For more details see the vignette: \code{vignette("simpleReporter", "teal.reporter")}.
}
