% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadModule.R
\name{download_report_srv}
\alias{download_report_srv}
\title{Download report Server}
\usage{
download_report_srv(
  id,
  reporter,
  notification = TRUE,
  output_types = c("pdf document", "html document", "powerpoint presentation",
    "word document"),
  show_modal = shiny::reactive(NULL)
)
}
\arguments{
\item{id}{\code{character}}

\item{reporter}{\code{Reporter} instance.}

\item{notification}{logical whether to add shiny notification about the download process, by default \code{TRUE}.}

\item{output_types}{\code{character} vector with \code{rmarkdown} output types,
by default \code{c("pdf document", "html document", "powerpoint presentation", "word document")}.}

\item{show_modal}{\code{reactive} to trigger popup of the download modal}
}
\value{
\code{shiny::moduleServer}
}
\description{
server for downloading the Report. Part of the simple Reporter.

For more details see the vignette: \code{vignette("simpleReporter", "teal.reporter")}.
}
