% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_addcard_id}
\alias{extract_addcard_id}
\title{Extract Add Card Button id}
\usage{
extract_addcard_id(input)
}
\arguments{
\item{input}{shiny input}
}
\description{
extract Add Card Button id.
It is needed to know when trigger the reactivity cycle for the \code{ReportCard}.
}
\note{
has to be invoked inside the reactive call.
}
