% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_addcard_input}
\alias{extract_addcard_input}
\title{Extract Add Card Button input}
\usage{
extract_addcard_input(input)
}
\arguments{
\item{input}{shiny input, \code{reactivevalues}.}
}
\description{
extract Add Card Button input.
It is needed to know when trigger the reactivity cycle for the \code{ReportCard}.
}
\note{
has to be invoked inside the reactive call.
}
