% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_bs_version}
\alias{get_bs_version}
\title{Get bootstrap current version}
\usage{
get_bs_version()
}
\description{
Get bootstrap current version
}
\note{
will work properly mainly inside a tag \code{.renderHook}
}
\keyword{internal}
