% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_utils.R
\name{get_yaml_field}
\alias{get_yaml_field}
\title{Extract field from \code{yaml} text}
\usage{
get_yaml_field(yaml_text, field_name)
}
\arguments{
\item{yaml_text}{(\code{rmd_yaml_header} or \code{character}) vector containing the \code{yaml} text.}

\item{field_name}{(\code{character}) the name of the field to extract.}
}
\value{
If the field is a list, it returns the names of elements in the list; otherwise,
it returns the extracted field.
}
\description{
Parses \code{yaml} text, extracting the specified field. Returns list names if it's a list;
otherwise, the field itself.
}
\keyword{internal}
