% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Archiver.R
\name{JSONArchiver}
\alias{JSONArchiver}
\title{\code{JSONArchiver}}
\description{
\code{JSONArchiver}

\code{JSONArchiver}
}
\keyword{internal}
\section{Super classes}{
\code{\link[teal.reporter:Archiver]{teal.reporter::Archiver}} -> \code{\link[teal.reporter:FileArchiver]{teal.reporter::FileArchiver}} -> \code{JSONArchiver}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-JSONArchiver-write}{\code{JSONArchiver$write()}}
\item \href{#method-JSONArchiver-read}{\code{JSONArchiver$read()}}
\item \href{#method-JSONArchiver-clone}{\code{JSONArchiver$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="teal.reporter" data-topic="FileArchiver" data-id="finalize"><a href='../../teal.reporter/html/FileArchiver.html#method-FileArchiver-finalize'><code>teal.reporter::FileArchiver$finalize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.reporter" data-topic="FileArchiver" data-id="get_output_dir"><a href='../../teal.reporter/html/FileArchiver.html#method-FileArchiver-get_output_dir'><code>teal.reporter::FileArchiver$get_output_dir()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.reporter" data-topic="FileArchiver" data-id="initialize"><a href='../../teal.reporter/html/FileArchiver.html#method-FileArchiver-initialize'><code>teal.reporter::FileArchiver$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JSONArchiver-write"></a>}}
\if{latex}{\out{\hypertarget{method-JSONArchiver-write}{}}}
\subsection{Method \code{write()}}{
write a \code{Reporter} instance in to this \code{JSONArchiver} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JSONArchiver$write(reporter)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reporter}}{\code{Reporter} instance.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisibly self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JSONArchiver-read"></a>}}
\if{latex}{\out{\hypertarget{method-JSONArchiver-read}{}}}
\subsection{Method \code{read()}}{
read a \code{Reporter} instance from a directory with \code{JSONArchiver}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JSONArchiver$read(path = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{\code{character(1)} a path to the directory with all proper files.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{Reporter} instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JSONArchiver-clone"></a>}}
\if{latex}{\out{\hypertarget{method-JSONArchiver-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JSONArchiver$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
