% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportDocument.R
\name{keep_in_report}
\alias{keep_in_report}
\title{Keep Objects In Report}
\usage{
keep_in_report(object, keep = TRUE)
}
\description{
Utility function to change behavior of \code{report_document()} elements to be
kept (\code{keep = TRUE}) or discarded (keep = \code{FALSE}) from the final \code{.Rmd} file containing downloaded report.
}
\details{
By default all R objects are only printed in the output document, but not kept in the \code{.Rmd} report.
By defaulf all text elements and \code{code_chunk} objects are kep both in the output document and \code{.Rmd} report.
}
