% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadReporterModule.R
\name{load_report_button}
\alias{load_report_button}
\alias{report_load_ui}
\alias{report_load_srv}
\title{Load \code{Reporter} button module}
\usage{
report_load_ui(id, label = NULL)

report_load_srv(id, reporter)
}
\arguments{
\item{id}{\code{character(1)} this \code{shiny} module's id.}

\item{label}{(\code{character(1)}) label of the button. By default it is empty.}

\item{reporter}{\code{\link{Reporter}} instance.}
}
\value{
\code{shiny::tagList}

\code{shiny::moduleServer}
}
\description{
Provides a button to upload \code{ReporterCard}(s) to the \code{Reporter}.

For more information, refer to the vignette: \code{vignette("simpleReporter", "teal.reporter")}.
}
