% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_utils.R
\name{md_header}
\alias{md_header}
\title{wrap a \code{yaml} string to the \code{markdown} header}
\usage{
md_header(x)
}
\arguments{
\item{x}{\code{character} \code{yaml} formatted string.}
}
\description{
wrap a \code{yaml} string to the \code{markdown} header.
}
\keyword{internal}
