% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{metadata<-}
\alias{metadata<-}
\alias{metadata<-.teal_card}
\alias{metadata<-.ReportCard}
\title{Set metadata for a \code{teal_card} or \code{ReportCard}}
\usage{
metadata(object, which) <- value

\method{metadata}{teal_card}(object, which) <- value

\method{metadata}{ReportCard}(object, which) <- value
}
\arguments{
\item{object}{(\code{teal_card} or \code{ReportCard}) The object to modify.}

\item{which}{(\code{character}) The name of the metadata field to set.}

\item{value}{The value to assign to the specified metadata field.}
}
\value{
The modified object with updated metadata.
}
\description{
This function allows you to set or modify metadata fields in a \code{teal_card} or \code{ReportCard} object.
It can be used to add new metadata or update existing fields.
}
\details{
The \code{ReportCard} class only supports the \code{title} field in metadata.
}
