% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportDocument.R
\name{metadata}
\alias{metadata}
\alias{metadata.ReportDocument}
\alias{metadata.ReportCard}
\title{Access metadata from a \code{ReportDocument} or \code{ReportCard}}
\usage{
metadata(object, which = NULL)

\method{metadata}{ReportDocument}(object, which = NULL)

\method{metadata}{ReportCard}(object, which = NULL)
}
\arguments{
\item{object}{(\code{ReportDocument} or \code{ReportCard}) The object from which to extract metadata.}

\item{which}{(\code{character} or \code{NULL}) The name of the metadata field to extract.}
}
\value{
A list of metadata fields or a specific field if \code{which} is provided.
}
\description{
This function retrieves metadata from a \code{ReportDocument} or \code{ReportCard} object.
When \code{which} is \code{NULL}, it returns all metadata fields as a list.
}
