% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editor_card.R
\name{module_editor_card}
\alias{module_editor_card}
\alias{ui_editor_card}
\alias{srv_editor_card}
\alias{ui_editor_card.default}
\title{Card editor module}
\usage{
ui_editor_card(id, value, cached_html = NULL)

srv_editor_card(id, card_r)

\method{ui_editor_card}{default}(id, value, cached_html)
}
\arguments{
\item{id}{(\code{character(1)}) A unique identifier for the module.}

\item{value}{(\verb{named list}) The content of the blocks to be edited. It can be a character string or other types.}

\item{cached_html}{(\verb{named list} of \code{shiny.tag} or \code{shiny.tag.list}) Cached HTML content to display in the UI.}
}
\value{
\itemize{
\item \code{ui_editor_card} returns \code{shiny.tag}
\item \code{srv_editor_card} returns \code{reactivevalues} named after name of the block element
}
}
\description{
Card editor module
}
