% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/previewer_content.R
\name{module_reporter_previewer_content}
\alias{module_reporter_previewer_content}
\alias{ui_reporter_previewer_content}
\alias{srv_reporter_previewer_content}
\alias{ui_reporter_previewer_content.default}
\alias{srv_reporter_previewer_content.default}
\title{Reporter previewer content module}
\usage{
ui_reporter_previewer_content(id, cached_content = rlang::list2())

srv_reporter_previewer_content(id, reporter)

\method{ui_reporter_previewer_content}{default}(id, cached_content = rlang::list2())

\method{srv_reporter_previewer_content}{default}(id, reporter)
}
\arguments{
\item{id}{(\code{character(1)}) A unique identifier for the module.}

\item{cached_content}{(\code{list})}

\item{reporter}{(\code{Reporter})}
}
\description{
Reporter previewer content module
}
