% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewpageBlock.R
\name{NewpageBlock}
\alias{NewpageBlock}
\title{\code{NewpageBlock}}
\description{
\code{NewpageBlock}

\code{NewpageBlock}
}
\examples{

## ------------------------------------------------
## Method `NewpageBlock$new`
## ------------------------------------------------

block <- teal.reporter:::NewpageBlock$new()

}
\keyword{internal}
\section{Super class}{
\code{\link[teal.reporter:ContentBlock]{teal.reporter::ContentBlock}} -> \code{NewpageBlock}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NewpageBlock$new()}}
\item \href{#method-clone}{\code{NewpageBlock$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="teal.reporter" data-topic="ContentBlock" data-id="deep_clone">}\href{../../teal.reporter/html/ContentBlock.html#method-deep_clone}{\code{teal.reporter::ContentBlock$deep_clone()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.reporter" data-topic="ContentBlock" data-id="get_content">}\href{../../teal.reporter/html/ContentBlock.html#method-get_content}{\code{teal.reporter::ContentBlock$get_content()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.reporter" data-topic="ContentBlock" data-id="set_content">}\href{../../teal.reporter/html/ContentBlock.html#method-set_content}{\code{teal.reporter::ContentBlock$set_content()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Returns a \code{NewpageBlock} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NewpageBlock$new()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Returns a \code{NewpageBlock} object with no content and the default style.
}

\subsection{Returns}{
\code{NewpageBlock}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{block <- teal.reporter:::NewpageBlock$new()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NewpageBlock$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
