% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{panel_item}
\alias{panel_item}
\title{Panel group widget}
\usage{
panel_item(title, ..., collapsed = TRUE, input_id = NULL)
}
\arguments{
\item{title}{(\code{character})\cr title of panel}

\item{...}{content of panel}

\item{collapsed}{(\code{logical}, optional)\cr
whether to initially collapse panel}

\item{input_id}{(\code{character}, optional)\cr
name of the panel item element. If supplied, this will register a shiny input variable that
indicates whether the panel item is open or collapsed and is accessed with \code{input$input_id}.}
}
\value{
(\code{shiny.tag})
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\keyword{internal}
