% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_utils.R
\name{print.yaml_header}
\alias{print.yaml_header}
\title{Print method for the \code{yaml_header} class}
\usage{
\method{print}{yaml_header}(x, ...)
}
\arguments{
\item{x}{\code{yaml_header} class object.}

\item{...}{optional text.}
}
\description{
Print method for the \code{yaml_header} class.
}
\examples{
input <- list(author = "", output = "pdf_document", toc = TRUE, keep_tex = TRUE)
out <- as_yaml_auto(input)
out
print(out)

}
