% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_card.R
\name{pseudo_code_chunk}
\alias{pseudo_code_chunk}
\title{Element of a report that should be treated as a code chunk, but can take other values}
\usage{
pseudo_code_chunk(object)
}
\arguments{
\item{object}{An object to be wrapped as a \code{pseudo_code_chunk}.}
}
\value{
An object of class \code{pseudo_code_chunk}.
}
\description{
This function creates a \code{pseudo_code_chunk} object, which is similar to a \code{code_chunk}
but can contain non-code elements. It is used to represent elements in a report
that are removed alongside with code chunks during certain processing steps.

For example, it can be used for title sections that don't make sense without the
associated code chunk.
}
\examples{
pseudo_code_chunk("## Analysis Title")
}
