% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadModule.R
\name{render_and_download}
\alias{render_and_download}
\title{Render and Download the Document}
\usage{
render_and_download(reporter, input, file)
}
\arguments{
\item{reporter}{\code{Reporter} instance.}

\item{input}{shiny input.}
}
\description{
render and download the document
}
\keyword{internal}
