% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadModule.R
\name{render_report}
\alias{render_report}
\title{Render the Report}
\usage{
render_report(reporter, input, file = tempdir())
}
\arguments{
\item{reporter}{\code{Reporter} instance.}

\item{input}{\code{reactivevalues} shiny input.}

\item{file}{\code{character} where to copy the returned directory.}
}
\value{
\code{file} argument
}
\description{
render the report and zip the created directory.
}
\keyword{internal}
