% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Renderer.R
\name{Renderer}
\alias{Renderer}
\title{\code{Renderer}}
\description{
\code{Renderer}

\code{Renderer}
}
\examples{

## ------------------------------------------------
## Method `Renderer$new`
## ------------------------------------------------

renderer <- teal.reporter:::Renderer$new()

## ------------------------------------------------
## Method `Renderer$renderRmd`
## ------------------------------------------------

yaml_l <- list(
  author = teal.reporter:::yaml_quoted("NEST"),
  title = teal.reporter:::yaml_quoted("Report"),
  date = teal.reporter:::yaml_quoted("07/04/2019"),
  output = list(pdf_document = list(keep_tex = TRUE))
)
yaml_header <- teal.reporter:::md_header(yaml::as.yaml(yaml_l))
text_block1 <- teal.reporter:::TextBlock$new()$set_content("text")$set_style("header2")
text_block2 <- teal.reporter:::TextBlock$new()$set_content("text")
text_block3 <- teal.reporter:::TextBlock$new()$set_content("tables")$set_style("header2")
ggplot1 <- ggplot2::ggplot(airquality, ggplot2::aes(Ozone, Month)) + ggplot2::geom_point()
picture_block <- teal.reporter:::PictureBlock$new()$set_content(ggplot1)
table_block1 <- teal.reporter:::TableBlock$new(iris)
lyt <- rtables::analyze(rtables::split_rows_by(rtables::basic_table(), "Day"), "Ozone", afun = mean)
table_res2 <- rtables::build_table(lyt, airquality)
table_block2 <- teal.reporter:::TableBlock$new(table_res2)
newpage_block <- teal.reporter:::NewpageBlock$new()
blocks <- list(
  text_block1, text_block2, picture_block, newpage_block,
  text_block3, table_block1, table_block2, newpage_block
)
renderer <- teal.reporter:::Renderer$new()
rmd_path <- renderer$renderRmd(blocks, yaml_header)

## ------------------------------------------------
## Method `Renderer$render`
## ------------------------------------------------

yaml_l <- list(
  author = teal.reporter:::yaml_quoted("NEST"),
  title = teal.reporter:::yaml_quoted("Report"),
  date = teal.reporter:::yaml_quoted("07/04/2019"),
  output = list(pdf_document = list(keep_tex = TRUE))
)
yaml_header <- teal.reporter:::md_header(yaml::as.yaml(yaml_l))
text_block1 <- teal.reporter:::TextBlock$new()$set_content("text")$set_style("header2")
text_block2 <- teal.reporter:::TextBlock$new()$set_content("text")
text_block3 <- teal.reporter:::TextBlock$new()$set_content("tables")$set_style("header2")
ggplot1 <- ggplot2::ggplot(airquality, ggplot2::aes(Ozone, Month)) + ggplot2::geom_point()
picture_block <- teal.reporter:::PictureBlock$new()$set_content(ggplot1)
table_block1 <- teal.reporter:::TableBlock$new(iris)
lyt <- rtables::analyze(rtables::split_rows_by(rtables::basic_table(), "Day"), "Ozone", afun = mean)
table_res2 <- rtables::build_table(lyt, airquality)
table_block2 <- teal.reporter:::TableBlock$new(table_res2)
newpage_block <- teal.reporter:::NewpageBlock$new()
blocks <- list(
  text_block1, text_block2, picture_block, newpage_block,
  text_block3, table_block1, table_block2, newpage_block
)
renderer <- teal.reporter:::Renderer$new()
result_path <- renderer$render(blocks, yaml_header)

## ------------------------------------------------
## Method `Renderer$get_output_dir`
## ------------------------------------------------

renderer <- teal.reporter:::Renderer$new()
renderer$get_output_dir()
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Renderer$new()}}
\item \href{#method-finalize}{\code{Renderer$finalize()}}
\item \href{#method-renderRmd}{\code{Renderer$renderRmd()}}
\item \href{#method-render}{\code{Renderer$render()}}
\item \href{#method-get_output_dir}{\code{Renderer$get_output_dir()}}
\item \href{#method-clone}{\code{Renderer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Returns a \code{Renderer} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Renderer$new()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Returns a \code{Renderer} object.
}

\subsection{Returns}{
\code{Renderer} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{renderer <- teal.reporter:::Renderer$new()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
finalize a \code{Renderer} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Renderer$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-renderRmd"></a>}}
\if{latex}{\out{\hypertarget{method-renderRmd}{}}}
\subsection{Method \code{renderRmd()}}{
getting the \code{Rmd} text which could be easily rendered later.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Renderer$renderRmd(blocks, yaml_header)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{blocks}}{\code{list} of \code{c("TextBlock", "PictureBlock", "NewpageBlock")} objects.}

\item{\code{yaml_header}}{\code{character} a \code{rmarkdown} \code{yaml} header.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{character} a \code{Rmd} text (\code{yaml} header + body), ready to be rendered.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{yaml_l <- list(
  author = teal.reporter:::yaml_quoted("NEST"),
  title = teal.reporter:::yaml_quoted("Report"),
  date = teal.reporter:::yaml_quoted("07/04/2019"),
  output = list(pdf_document = list(keep_tex = TRUE))
)
yaml_header <- teal.reporter:::md_header(yaml::as.yaml(yaml_l))
text_block1 <- teal.reporter:::TextBlock$new()$set_content("text")$set_style("header2")
text_block2 <- teal.reporter:::TextBlock$new()$set_content("text")
text_block3 <- teal.reporter:::TextBlock$new()$set_content("tables")$set_style("header2")
ggplot1 <- ggplot2::ggplot(airquality, ggplot2::aes(Ozone, Month)) + ggplot2::geom_point()
picture_block <- teal.reporter:::PictureBlock$new()$set_content(ggplot1)
table_block1 <- teal.reporter:::TableBlock$new(iris)
lyt <- rtables::analyze(rtables::split_rows_by(rtables::basic_table(), "Day"), "Ozone", afun = mean)
table_res2 <- rtables::build_table(lyt, airquality)
table_block2 <- teal.reporter:::TableBlock$new(table_res2)
newpage_block <- teal.reporter:::NewpageBlock$new()
blocks <- list(
  text_block1, text_block2, picture_block, newpage_block,
  text_block3, table_block1, table_block2, newpage_block
)
renderer <- teal.reporter:::Renderer$new()
rmd_path <- renderer$renderRmd(blocks, yaml_header)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render"></a>}}
\if{latex}{\out{\hypertarget{method-render}{}}}
\subsection{Method \code{render()}}{
render blocks and \code{yaml} header to the output file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Renderer$render(blocks, yaml_header, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{blocks}}{\code{list} of \code{c("TextBlock", "PictureBlock", "NewpageBlock")} objects.}

\item{\code{yaml_header}}{\code{character} a \code{rmarkdown} \code{yaml} header.}

\item{\code{...}}{\code{rmarkdown::render} arguments, \code{input} and \code{output_dir} should not be updated.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{character} path to the output
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{yaml_l <- list(
  author = teal.reporter:::yaml_quoted("NEST"),
  title = teal.reporter:::yaml_quoted("Report"),
  date = teal.reporter:::yaml_quoted("07/04/2019"),
  output = list(pdf_document = list(keep_tex = TRUE))
)
yaml_header <- teal.reporter:::md_header(yaml::as.yaml(yaml_l))
text_block1 <- teal.reporter:::TextBlock$new()$set_content("text")$set_style("header2")
text_block2 <- teal.reporter:::TextBlock$new()$set_content("text")
text_block3 <- teal.reporter:::TextBlock$new()$set_content("tables")$set_style("header2")
ggplot1 <- ggplot2::ggplot(airquality, ggplot2::aes(Ozone, Month)) + ggplot2::geom_point()
picture_block <- teal.reporter:::PictureBlock$new()$set_content(ggplot1)
table_block1 <- teal.reporter:::TableBlock$new(iris)
lyt <- rtables::analyze(rtables::split_rows_by(rtables::basic_table(), "Day"), "Ozone", afun = mean)
table_res2 <- rtables::build_table(lyt, airquality)
table_block2 <- teal.reporter:::TableBlock$new(table_res2)
newpage_block <- teal.reporter:::NewpageBlock$new()
blocks <- list(
  text_block1, text_block2, picture_block, newpage_block,
  text_block3, table_block1, table_block2, newpage_block
)
renderer <- teal.reporter:::Renderer$new()
result_path <- renderer$render(blocks, yaml_header)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_output_dir"></a>}}
\if{latex}{\out{\hypertarget{method-get_output_dir}{}}}
\subsection{Method \code{get_output_dir()}}{
get \code{output_dir} field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Renderer$get_output_dir()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} a \code{output_dir} field path.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{renderer <- teal.reporter:::Renderer$new()
renderer$get_output_dir()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Renderer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
