% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_report-report.R
\name{report<-}
\alias{report<-}
\title{Replace a report in \code{teal_report}}
\usage{
report(x) <- value
}
\arguments{
\item{x}{(\code{teal_report})}

\item{value}{(\code{doc})}
}
\value{
\code{teal_report}
}
\description{
Replace a report in \code{teal_report}
}
