% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_reportable-report.R
\name{report}
\alias{report}
\title{Extract report from \code{teal_reportable}}
\usage{
report(x)
}
\arguments{
\item{x}{(\code{teal_reportable})}
}
\value{
\code{teal_reportable}
}
\description{
Extract report from \code{teal_reportable}
}
