% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportDocument.R
\name{report_document}
\alias{report_document}
\alias{ReportDocument}
\alias{c.ReportDocument}
\alias{[.ReportDocument}
\title{\code{ReportDocument}: An \code{S3} class for managing \code{teal} reports}
\usage{
report_document()

\method{c}{ReportDocument}(...)

\method{[}{ReportDocument}(x, i)
}
\arguments{
\item{...}{objects passed to \code{c()} function}

\item{x}{\code{ReportDocument} object}

\item{values}{objects to be included in the modified \code{ReportDocument}}
}
\value{
An \code{S3} \code{list} of class \code{ReportDocument}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This \code{S3} class is designed to store, manage, edit and adjust report cards.
It facilitates the creation, manipulation, and serialization of report-related data.
}
\examples{
report <- report_document()
class(report)
report <- c(report, list("## Headline"), list("## Table"), list(summary(iris)))
report <- report[1:2]
report <- append(report, c(list("## Table 2"), list(summary(mtcars))), after = 1)
class(report)

}
