% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadReporterModule.R
\name{report_load_srv}
\alias{report_load_srv}
\title{Server to Load \code{Reporter}}
\usage{
report_load_srv(id, reporter)
}
\arguments{
\item{id}{\code{character(1)} this \code{shiny} module's id.}

\item{reporter}{\code{\link{Reporter}} instance.}
}
\value{
\code{shiny::moduleServer}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Server to load \code{ReporterCard}(s) to the \code{Reporter}

For more details see the vignette: \code{vignette("simpleReporter", "teal.reporter")}.
}
