% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadReporterModule.R
\name{report_load_ui}
\alias{report_load_ui}
\title{Load Report User Interface}
\usage{
report_load_ui(id)
}
\arguments{
\item{id}{\code{character(1)} this \code{shiny} module's id.}
}
\value{
\code{shiny::tagList}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
button for adding views/cards to the Report.

For more details see the vignette: \code{vignette("simpleReporter", "teal.reporter")}.
}
