% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadModule.R
\name{report_render_and_compress}
\alias{report_render_and_compress}
\title{Render the Report}
\usage{
report_render_and_compress(reporter, input, file = tempdir())
}
\arguments{
\item{reporter}{\code{Reporter} instance.}

\item{input}{\code{list} like shiny input converted to a regular list.}

\item{file}{\code{character} where to copy the returned directory.}
}
\value{
\code{file} argument
}
\description{
render the report and zip the created directory.
}
\keyword{internal}
