% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Previewer.R
\name{reporter_previewer}
\alias{reporter_previewer}
\alias{preview_report_button_ui}
\alias{preview_report_button_srv}
\title{Report previewer module}
\usage{
preview_report_button_ui(id, label = "Preview Report")

preview_report_button_srv(id, reporter)
}
\arguments{
\item{id}{(\code{character(1)}) \code{shiny} module instance id.}

\item{label}{(\code{character(1)}) label of the button. By default it is "Preview Report".}

\item{reporter}{(\code{Reporter}) instance.}
}
\value{
\code{NULL}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Module offers functionalities to visualize, manipulate,
and interact with report cards that have been added to a report.
It includes a previewer interface to see the cards and options to modify the report before downloading.

Cards are saved by the \code{shiny} bookmarking mechanism.

For more details see the vignette: \code{vignette("previewerReporter", "teal.reporter")}.
}
\details{
To access the default values for the \code{global_knitr} parameter,
use \code{getOption('teal.reporter.global_knitr')}. These defaults include:
\itemize{
\item \code{echo = TRUE}
\item \code{tidy.opts = list(width.cutoff = 60)}
\item \code{tidy = TRUE}  if \code{formatR} package is installed, \code{FALSE} otherwise
}
}
