% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Previewer.R
\name{reporter_previewer_srv}
\alias{reporter_previewer_srv}
\title{Reporter Previewer Server}
\usage{
reporter_previewer_srv(
  id,
  reporter,
  rmd_yaml_args = list(author = "NEST", title = "Report", date =
    as.character(Sys.Date()), output = "html_document")
)
}
\arguments{
\item{id}{\code{character}}

\item{reporter}{\code{Reporter} instance}

\item{rmd_yaml_args}{\verb{named list} vector with \code{Rmd} \code{yaml} header fields and their default values.
Default \code{list(author = "NEST", title = "Report", date = Sys.Date(), output = "html_document")}.
Please update only values at this moment.}
}
\description{
server supporting the functionalities of the reporter previewer
}
