% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Previewer.R
\name{reporter_previewer_ui}
\alias{reporter_previewer_ui}
\title{Reporter Previewer User Interface}
\usage{
reporter_previewer_ui(id)
}
\arguments{
\item{id}{\code{character(1)} this \code{shiny} module's id.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
reporter previewer user interface to visualize and manipulate the already added report Cards
}
