% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResetModule.R
\name{reset_report_button}
\alias{reset_report_button}
\alias{reset_report_button_ui}
\alias{reset_report_button_srv}
\title{Reset report button module}
\usage{
reset_report_button_ui(id, label = "Reset Report")

reset_report_button_srv(id, reporter)
}
\arguments{
\item{id}{(\code{character(1)}) \code{shiny} module instance id.}

\item{label}{(\code{character(1)}) label of the button. By default \verb{Reset Report}.}

\item{reporter}{(\code{Reporter}) instance.}
}
\value{
\code{NULL}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Provides a button that triggers resetting the report content.

For more information, refer to the vignette: \code{vignette("simpleReporter", "teal.reporter")}.
}
