% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResetModule.R
\name{reset_report_button_srv}
\alias{reset_report_button_srv}
\title{Reset Button Server}
\usage{
reset_report_button_srv(id, reporter)
}
\arguments{
\item{id}{\code{character(1)} \code{shiny} module id.}

\item{reporter}{\code{\link{Reporter}} instance.}
}
\value{
\code{shiny::moduleServer}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
server for resetting the Report content.

For more details see the vignette: \code{vignette("simpleReporter", "teal.reporter")}.
}
