% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResetModule.R
\name{reset_report_button_ui}
\alias{reset_report_button_ui}
\title{Reset Button Reporter User Interface}
\usage{
reset_report_button_ui(id)
}
\arguments{
\item{id}{\code{character}}
}
\value{
\code{shiny::tagList}
}
\description{
button for reseting the report content.

For more details see the vignette: \code{vignette("simpleReporter", "teal.reporter")}.
}
