% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_utils.R
\name{reverse_yaml_field}
\alias{reverse_yaml_field}
\title{Extract a Field from \code{yaml} and Optionally Retrieve Names from a List}
\usage{
reverse_yaml_field(yaml_text, field_name)
}
\arguments{
\item{yaml_text}{A character vector containing the \code{yaml} text.}

\item{field_name}{The name of the field to extract.}
}
\value{
it returns the names of elements in the list; otherwise,
it returns the extracted field.
}
\description{
This function parses a \code{YAML} text and extracts the specified field. It provides
the option to retrieve the names of elements from a list if the field contains a list.
}
\keyword{internal}
