% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_utils.R
\name{rmd_output_arguments}
\alias{rmd_output_arguments}
\title{Get document output arguments from the \code{rmarkdown} package}
\usage{
rmd_output_arguments(output_name, default_values = FALSE)
}
\arguments{
\item{output_name}{\verb{character``rmarkdown} output name.}

\item{default_values}{\code{logical} if to return a default values for each argument.}
}
\description{
get document output arguments from the \code{rmarkdown} package
}
\examples{
rmd_output_arguments("pdf_document")
rmd_output_arguments("pdf_document", TRUE)

}
