% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_utils.R
\name{rmd_outputs}
\alias{rmd_outputs}
\title{Get document output types from the \code{rmarkdown} package}
\usage{
rmd_outputs()
}
\value{
\code{character} vector.
}
\description{
Retrieves vector of available document output types from the \code{rmarkdown} package,
such as \code{pdf_document}, \code{html_document}, etc.
}
\examples{
rmd_outputs()
}
