% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleReporter.R
\name{simple_reporter_ui}
\alias{simple_reporter_ui}
\title{Simple Reporter user interface}
\usage{
simple_reporter_ui(id)
}
\arguments{
\item{id}{character}
}
\value{
shiny \code{tagList}
}
\description{
two buttons for adding views and downloading the Report.
}
