% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_text_block}
\alias{split_text_block}
\title{Divide text block into smaller blocks}
\usage{
split_text_block(x, n)
}
\arguments{
\item{n}{(\code{integer}) number of lines per block}

\item{block_text}{(\code{character}) string containing the input block of text}
}
\value{
List of character strings with up to \code{n} lines in each element.
}
\description{
Split a text block into smaller blocks with a specified number of lines.
}
\details{
A single character string containing a text block of multiple lines (separated by \verb{\\n})
is split into multiple strings with n or less lines each.
}
\keyword{internal}
