% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableBlock.R
\name{TableBlock}
\alias{TableBlock}
\title{\code{TableBlock}}
\description{
\code{TableBlock}

\code{TableBlock}
}
\examples{

## ------------------------------------------------
## Method `TableBlock$new`
## ------------------------------------------------

block <- teal.reporter:::TableBlock$new()


## ------------------------------------------------
## Method `TableBlock$set_content`
## ------------------------------------------------

block <- teal.reporter:::TableBlock$new()
table_path <- tempfile(fileext = ".csv")
write.csv(mtcars, table_path)
block$set_content(table_path)


## ------------------------------------------------
## Method `TableBlock$get_content`
## ------------------------------------------------

block <- teal.reporter:::TableBlock$new()
block$get_content()


## ------------------------------------------------
## Method `TableBlock$get_type`
## ------------------------------------------------

block <- teal.reporter:::TableBlock$new()
block$get_type()

}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TableBlock$new()}}
\item \href{#method-set_content}{\code{TableBlock$set_content()}}
\item \href{#method-get_content}{\code{TableBlock$get_content()}}
\item \href{#method-get_type}{\code{TableBlock$get_type()}}
\item \href{#method-clone}{\code{TableBlock$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Returns a \code{TableBlock} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableBlock$new()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Returns a \code{TableBlock} object with no content and the default style.
}

\subsection{Returns}{
\code{TableBlock}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{block <- teal.reporter:::TableBlock$new()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_content"></a>}}
\if{latex}{\out{\hypertarget{method-set_content}{}}}
\subsection{Method \code{set_content()}}{
Sets the content of this \code{TableBlock}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableBlock$set_content(content)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{content}}{(\code{character(1)}) a string path assigned to this \code{TableBlock}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
throws if argument is not \code{character(1)}.
}

\subsection{Returns}{
invisibly self
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{block <- teal.reporter:::TableBlock$new()
table_path <- tempfile(fileext = ".csv")
write.csv(mtcars, table_path)
block$set_content(table_path)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_content"></a>}}
\if{latex}{\out{\hypertarget{method-get_content}{}}}
\subsection{Method \code{get_content()}}{
Returns the content of this \code{TableBlock}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableBlock$get_content()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the content of this \code{TableBlock}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{block <- teal.reporter:::TableBlock$new()
block$get_content()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_type"></a>}}
\if{latex}{\out{\hypertarget{method-get_type}{}}}
\subsection{Method \code{get_type()}}{
Returns the type of this \code{TableBlock}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableBlock$get_type()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the type of this \code{TableBlock}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{block <- teal.reporter:::TableBlock$new()
block$get_type()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableBlock$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
