% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_report-class.R
\name{teal_report}
\alias{teal_report}
\alias{as.teal_report}
\title{Comprehensive data integration function for \code{teal} applications}
\usage{
teal_report(
  ...,
  card = NULL,
  code = character(0),
  join_keys = teal.data::join_keys()
)

as.teal_report(x)
}
\arguments{
\item{card}{(\code{card}) object containing the report content.}

\item{x}{(\code{qenv} or \code{teal_data}) object to convert to \code{teal_report}.}
}
\value{
A \code{teal_report} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Initializes a reportable data for \code{teal} application.
}
\examples{
teal_report(x1 = iris, x2 = mtcars)
}
\seealso{
\code{\link[teal.data:teal_data]{teal.data::teal_data}}
}
