% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_reportable-class.R
\name{teal_reportable}
\alias{teal_reportable}
\title{Comprehensive data integration function for \code{teal} applications}
\usage{
teal_reportable(
  ...,
  report = report_document(),
  code = character(0),
  join_keys = teal.data::join_keys()
)
}
\arguments{
\item{...}{any number of objects (presumably data objects) provided as \code{name = value} pairs.}

\item{code}{(\code{character}, \code{language}) optional code to reproduce the datasets provided in \code{...}.
Note this code is not executed and the \code{teal_data} may not be reproducible

Use \code{\link[teal.data:verify]{verify()}} to verify code reproducibility.}

\item{join_keys}{(\code{join_keys} or single \code{join_key_set})
optional object with datasets column names used for joining.
If empty then no joins between pairs of objects.}

\item{raport}{(\code{doc})}
}
\value{
A \code{teal_reportable} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Initializes a reportable data for \code{teal} application.
}
\examples{
teal_reportable(x1 = iris, x2 = mtcars)

}
\seealso{
\code{\link[teal.data:teal_data]{teal.data::teal_data}}
}
