% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextBlock.R
\name{TextBlock}
\alias{TextBlock}
\title{\code{TextBlock}}
\description{
\code{TextBlock}

\code{TextBlock}
}
\examples{

## ------------------------------------------------
## Method `TextBlock$new`
## ------------------------------------------------

block <- teal.reporter:::TextBlock$new()


## ------------------------------------------------
## Method `TextBlock$set_style`
## ------------------------------------------------

block <- teal.reporter:::TextBlock$new()
block$set_style("header")


## ------------------------------------------------
## Method `TextBlock$get_style`
## ------------------------------------------------

block <- teal.reporter:::TextBlock$new()
block$get_style()


## ------------------------------------------------
## Method `TextBlock$set_content`
## ------------------------------------------------

block <- teal.reporter:::TextBlock$new()
block$set_content("Some text")


## ------------------------------------------------
## Method `TextBlock$get_content`
## ------------------------------------------------

block <- teal.reporter:::TextBlock$new()
block$get_content()


## ------------------------------------------------
## Method `TextBlock$get_available_styles`
## ------------------------------------------------

block <- teal.reporter:::TextBlock$new()
block$get_available_styles()

}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TextBlock$new()}}
\item \href{#method-set_style}{\code{TextBlock$set_style()}}
\item \href{#method-get_style}{\code{TextBlock$get_style()}}
\item \href{#method-set_content}{\code{TextBlock$set_content()}}
\item \href{#method-get_content}{\code{TextBlock$get_content()}}
\item \href{#method-get_available_styles}{\code{TextBlock$get_available_styles()}}
\item \href{#method-clone}{\code{TextBlock$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Returns a \code{TextBlock} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextBlock$new()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Returns a \code{TextBlock} object with no content and the default style.
}

\subsection{Returns}{
\code{TextBlock}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{block <- teal.reporter:::TextBlock$new()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_style"></a>}}
\if{latex}{\out{\hypertarget{method-set_style}{}}}
\subsection{Method \code{set_style()}}{
Sets the style of this \code{TextBlock}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextBlock$set_style(style)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{style}}{(\code{character(1)}) one of: \code{default}, \code{header}, \code{verbatim}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The style has bearing on the rendering of this block.
}

\subsection{Returns}{
invisibly self
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{block <- teal.reporter:::TextBlock$new()
block$set_style("header")

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_style"></a>}}
\if{latex}{\out{\hypertarget{method-get_style}{}}}
\subsection{Method \code{get_style()}}{
Returns the style of this \code{TextBlock}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextBlock$get_style()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character(1)} the style of this \code{TextBlock}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{block <- teal.reporter:::TextBlock$new()
block$get_style()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_content"></a>}}
\if{latex}{\out{\hypertarget{method-set_content}{}}}
\subsection{Method \code{set_content()}}{
Sets the content of this \code{TextBlock}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextBlock$set_content(content)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{content}}{(\code{character(1)}) a string assigned to this \code{TextBlock}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
throws if argument is not \code{character(1)}.
}

\subsection{Returns}{
invisibly self
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{block <- teal.reporter:::TextBlock$new()
block$set_content("Some text")

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_content"></a>}}
\if{latex}{\out{\hypertarget{method-get_content}{}}}
\subsection{Method \code{get_content()}}{
Returns the content of this \code{TextBlock}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextBlock$get_content()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the content of this \code{TextBlock}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{block <- teal.reporter:::TextBlock$new()
block$get_content()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_available_styles"></a>}}
\if{latex}{\out{\hypertarget{method-get_available_styles}{}}}
\subsection{Method \code{get_available_styles()}}{
Returns an array of styles available to this \code{TextBlock}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextBlock$get_available_styles()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a \code{character} array of styles
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{block <- teal.reporter:::TextBlock$new()
block$get_available_styles()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextBlock$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
