% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_flextable}
\alias{to_flextable}
\title{Convert content into a \code{flextable}}
\usage{
to_flextable(content)
}
\arguments{
\item{content}{Supported formats: \code{data.frame}, \code{rtables}, \code{TableTree}, \code{ElementaryTable}, \code{listing_df}}
}
\value{
\code{flextable}.
}
\description{
Converts supported table formats into a \code{flextable} for enhanced formatting and presentation.
}
\details{
Function merges cells with \code{colspan} > 1,
aligns columns to the center and row names to the left,
indents the row names by 10 times indentation.
}
\keyword{internal}
