% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Previewer.R
\name{toHTML.ContentBlock}
\alias{toHTML.ContentBlock}
\title{Convert a \code{ContentBlock} to HTML}
\usage{
\method{toHTML}{ContentBlock}(x, ...)
}
\arguments{
\item{x}{ An object to display. }

\item{...}{ Optional parameters for methods; the \code{"packageIQR"} and
  \code{"news_db"} methods pass these to \code{\link[tools]{HTMLheader}}. }
}
\description{
Convert a \code{ContentBlock} to HTML
}
\keyword{internal}
