% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddCardModule.R
\name{add_card_button_srv}
\alias{add_card_button_srv}
\title{Add Card Button Server}
\usage{
add_card_button_srv(id, reporter, card_fun)
}
\arguments{
\item{id}{\code{character}}

\item{reporter}{\code{Reporter} instance.}

\item{card_fun}{\code{function} which returns a \code{ReportCard} instance.}
}
\value{
\code{shiny::moduleServer}
}
\description{
server for adding views/cards the Report. Part of the simple Reporter.

For more details see the vignette: \code{vignette("simpleReporter", "teal.reporter")}.
}
