% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddCardModule.R
\name{add_card_button_ui}
\alias{add_card_button_ui}
\title{Add Card Reporter user interface}
\usage{
add_card_button_ui(id)
}
\arguments{
\item{id}{character}
}
\value{
shiny \code{tagList}
}
\description{
button for adding views/cards to the Report. Part of the simple Reporter user interface.
}
