% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddCardModule.R
\name{add_card_button_ui}
\alias{add_card_button_ui}
\title{Add Card Button User Interface}
\usage{
add_card_button_ui(id, label = NULL)
}
\arguments{
\item{id}{\code{character(1)} this \code{shiny} module's id.}

\item{label}{\code{character(1)} label before the icon, if used then dynamic hover label is not available.
By default \code{NULL} so a dynamic hover label is used.}
}
\value{
\code{shiny::tagList}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
button for adding views/cards to the Report.

For more details see the vignette: \code{vignette("simpleReporter", "teal.reporter")}.
}
