% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Archiver.R
\name{Archiver}
\alias{Archiver}
\title{\code{Archiver}}
\description{
\code{Archiver}

\code{Archiver}
}
\examples{

## ------------------------------------------------
## Method `Archiver$new`
## ------------------------------------------------

archiver <- getFromNamespace("Archiver", "teal.reporter")$new()
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Archiver-new}{\code{Archiver$new()}}
\item \href{#method-Archiver-finalize}{\code{Archiver$finalize()}}
\item \href{#method-Archiver-read}{\code{Archiver$read()}}
\item \href{#method-Archiver-write}{\code{Archiver$write()}}
\item \href{#method-Archiver-clone}{\code{Archiver$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Archiver-new"></a>}}
\if{latex}{\out{\hypertarget{method-Archiver-new}{}}}
\subsection{Method \code{new()}}{
Returns an \code{Archiver} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Archiver$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
an \code{Archiver} object
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{archiver <- getFromNamespace("Archiver", "teal.reporter")$new()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Archiver-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-Archiver-finalize}{}}}
\subsection{Method \code{finalize()}}{
Finalizes an \code{Archiver} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Archiver$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Archiver-read"></a>}}
\if{latex}{\out{\hypertarget{method-Archiver-read}{}}}
\subsection{Method \code{read()}}{
Pure virtual method for reading an \code{Archiver}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Archiver$read()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Archiver-write"></a>}}
\if{latex}{\out{\hypertarget{method-Archiver-write}{}}}
\subsection{Method \code{write()}}{
Pure virtual method for writing an \code{Archiver}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Archiver$write()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Archiver-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Archiver-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Archiver$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
