% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_report-c.R
\name{c.teal_report}
\alias{c.teal_report}
\title{Concatenate teal_report objects}
\usage{
\method{c}{teal_report}(...)
}
\arguments{
\item{...}{teal_report objects to concatenate}
}
\value{
A teal_report object with concatenated code, .xData, and teal_card fields
}
\description{
Concatenate teal_report objects
}
