% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_document.R
\name{code_chunk}
\alias{code_chunk}
\title{Generate an R Markdown code chunk}
\usage{
code_chunk(code, ...)
}
\arguments{
\item{code}{A character string containing the R code.}

\item{...}{Additional named parameters to be included as chunk options (e.g., \code{echo = TRUE}).}
}
\value{
An object of class \code{code_chunk}.
}
\description{
This function creates a \code{code_chunk} object, which represents an R Markdown
code chunk. It stores the R code and any specified chunk options (e.g., \code{echo}, \code{eval}).
These objects are typically processed later to generate the final R Markdown text.
}
\examples{
my_chunk <- code_chunk("x <- 1:10", echo = TRUE, message = FALSE)
class(my_chunk)
attributes(my_chunk)$param
}
