% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadModule.R
\name{download_report_button_srv}
\alias{download_report_button_srv}
\title{Download Button Server}
\usage{
download_report_button_srv(
  id,
  reporter,
  rmd_output = c(html = "html_document", pdf = "pdf_document", powerpoint =
    "powerpoint_presentation", word = "word_document"),
  rmd_yaml_args = list(author = "NEST", title = "Report", date =
    as.character(Sys.Date()), output = "html_document")
)
}
\arguments{
\item{id}{\code{character(1)} this \code{shiny} module's id.}

\item{reporter}{\code{\link{Reporter}} instance.}

\item{rmd_output}{\code{character} vector with \code{rmarkdown} output types,
by default all possible \code{c("pdf_document", "html_document", "powerpoint_presentation", "word_document")}.
\code{pdflatex} has to be installed to use the \code{"pdf_document"} output. If vector is named then the names
will appear in the \code{UI}.}

\item{rmd_yaml_args}{\verb{named list} vector with \code{Rmd} \code{yaml} header fields and their default values.
Default \code{list(author = "NEST", title = "Report", date = Sys.Date(), output = "html_document")}.
Please update only values at this moment.}
}
\value{
\code{shiny::moduleServer}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
server for downloading the Report.

For more details see the vignette: \code{vignette("simpleReporter", "teal.reporter")}.
}
