% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadModule.R
\name{download_report_button_srv}
\alias{download_report_button_srv}
\title{Download Button Server}
\usage{
download_report_button_srv(id, reporter, notification = TRUE)
}
\arguments{
\item{id}{character}

\item{reporter}{\code{Reporter} instance.}

\item{notification}{logical if to add shiny notification about the download process.}
}
\value{
shiny \code{moduleServer}
}
\description{
server for downloading the Report. Part of the simple Reporter.
}
