% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadModule.R
\name{download_report_button_srv}
\alias{download_report_button_srv}
\title{Download Button Server}
\usage{
download_report_button_srv(
  id,
  reporter,
  notification = TRUE,
  output_types = c("pdf_document", "html_document", "powerpoint_presentation",
    "word_document")
)
}
\arguments{
\item{id}{\code{character}}

\item{reporter}{\code{Reporter} instance.}

\item{notification}{\code{logical} whether to add a shiny notification about the download process. Default \code{TRUE}.}

\item{output_types}{\code{character} vector with \code{rmarkdown} output types,
by default \code{c("pdf_document", "html_document", "powerpoint_presentation", "word_document")}.}
}
\value{
\code{shiny::moduleServer}
}
\description{
server for downloading the Report.

For more details see the vignette: \code{vignette("simpleReporter", "teal.reporter")}.
}
