% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadModule.R
\name{download_report_button_srv}
\alias{download_report_button_srv}
\title{Download Button Server}
\usage{
download_report_button_srv(
  id,
  reporter,
  notification = TRUE,
  rmd_output = c("html_document", "pdf_document", "powerpoint_presentation",
    "word_document"),
  rmd_yaml_args = list(author = "NEST", title = "Report", date =
    as.character(Sys.Date()), output = "html_document")
)
}
\arguments{
\item{id}{\code{character}}

\item{reporter}{\code{Reporter} instance.}

\item{notification}{\code{logical} whether to add a shiny notification about the download process. Default \code{TRUE}.}

\item{rmd_output}{\code{character} vector with \code{rmarkdown} output types,
by default all possible \code{c("pdf_document", "html_document", "powerpoint_presentation", "word_document")}.}

\item{rmd_yaml_args}{\verb{named list} vector with \code{Rmd} \code{yaml} header fields and their default values.
Default \code{list(author = "NEST", title = "Report", date = Sys.Date(), output = "html_document")}.
Please update only values at this moment.}
}
\value{
\code{shiny::moduleServer}
}
\description{
server for downloading the Report.

For more details see the vignette: \code{vignette("simpleReporter", "teal.reporter")}.
}
