% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadModule.R
\name{download_report_button_srv}
\alias{download_report_button_srv}
\title{Download Button Server}
\usage{
download_report_button_srv(id, reporter, notification = TRUE)
}
\arguments{
\item{id}{\code{character}}

\item{reporter}{\code{Reporter} instance.}

\item{notification}{logical whether to add shiny notification about the download process.}
}
\value{
\code{shiny::tagList}
}
\description{
server for downloading the Report. Part of the simple Reporter.

For more details see the vignette: \code{vignette("simpleReporter", "teal.reporter")}.
}
