% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadModule.R
\name{download_report_button_ui}
\alias{download_report_button_ui}
\title{Download Button Reporter User Interface}
\usage{
download_report_button_ui(id, label = "")
}
\arguments{
\item{id}{\code{character(1)} this \code{shiny} module's id.}

\item{label}{\code{character(1)} label before the icon.}
}
\value{
\code{shiny::tagList}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
button for downloading the Report.

For more details see the vignette: \code{vignette("simpleReporter", "teal.reporter")}.
}
