% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadModule.R
\name{download_report_button_ui}
\alias{download_report_button_ui}
\title{Download Button Reporter UI}
\usage{
download_report_button_ui(id)
}
\arguments{
\item{id}{character}
}
\value{
shiny \code{tagList}
}
\description{
button for downloading the Report. Part of the simple Reporter UI.
}
