% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_addcard_input}
\alias{extract_addcard_input}
\title{Extract Add Card Button input}
\usage{
extract_addcard_input(input)
}
\arguments{
\item{input}{shiny input, \code{reactivevalues}.}
}
\value{
value
}
\description{
extract Add Card Button input, looks for an id with a suffix \code{addReportCardButton}.
It is needed to know when to trigger the reactivity cycle for the \code{ReportCard}.
}
\note{
has to be invoked inside the active reactive context.
}
