% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Archiver.R
\name{FileArchiver}
\alias{FileArchiver}
\title{\code{RDSArchiver}}
\description{
\code{RDSArchiver}

\code{RDSArchiver}
}
\examples{

## ------------------------------------------------
## Method `FileArchiver$new`
## ------------------------------------------------

archiver <- teal.reporter:::FileArchiver$new()


## ------------------------------------------------
## Method `FileArchiver$get_output_dir`
## ------------------------------------------------

archiver <- teal.reporter:::FileArchiver$new()
archiver$get_output_dir()
}
\keyword{internal}
\section{Super class}{
\code{\link[teal.reporter:Archiver]{teal.reporter::Archiver}} -> \code{RDSArchiver}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RDSArchiver-new}{\code{FileArchiver$new()}}
\item \href{#method-RDSArchiver-finalize}{\code{FileArchiver$finalize()}}
\item \href{#method-RDSArchiver-get_output_dir}{\code{FileArchiver$get_output_dir()}}
\item \href{#method-RDSArchiver-clone}{\code{FileArchiver$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="teal.reporter" data-topic="Archiver" data-id="read"><a href='../../teal.reporter/html/Archiver.html#method-Archiver-read'><code>teal.reporter::Archiver$read()</code></a></li>
<li><span class="pkg-link" data-pkg="teal.reporter" data-topic="Archiver" data-id="write"><a href='../../teal.reporter/html/Archiver.html#method-Archiver-write'><code>teal.reporter::Archiver$write()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RDSArchiver-new"></a>}}
\if{latex}{\out{\hypertarget{method-RDSArchiver-new}{}}}
\subsection{Method \code{new()}}{
Returns a \code{FileArchiver} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileArchiver$new(version)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a \code{FileArchiver} object
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{archiver <- teal.reporter:::FileArchiver$new()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RDSArchiver-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-RDSArchiver-finalize}{}}}
\subsection{Method \code{finalize()}}{
Finalizes a \code{FileArchiver} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileArchiver$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RDSArchiver-get_output_dir"></a>}}
\if{latex}{\out{\hypertarget{method-RDSArchiver-get_output_dir}{}}}
\subsection{Method \code{get_output_dir()}}{
get \code{output_dir} field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileArchiver$get_output_dir()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} a \code{output_dir} field path.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{archiver <- teal.reporter:::FileArchiver$new()
archiver$get_output_dir()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RDSArchiver-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RDSArchiver-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileArchiver$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
