% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileBlock.R
\name{FileBlock}
\alias{FileBlock}
\title{\code{FileBlock}}
\description{
\code{FileBlock}

\code{FileBlock}
}
\examples{

## ------------------------------------------------
## Method `FileBlock$from_list`
## ------------------------------------------------

block <- teal.reporter:::FileBlock$new()
file_path <- tempfile(fileext = ".png")
saveRDS(iris, file_path)
block$from_list(list(path = file_path))

## ------------------------------------------------
## Method `FileBlock$to_list`
## ------------------------------------------------

block <- teal.reporter:::FileBlock$new()
block$to_list()
}
\keyword{internal}
\section{Super class}{
\code{\link[teal.reporter:ContentBlock]{teal.reporter::ContentBlock}} -> \code{FileBlock}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FileBlock-finalize}{\code{FileBlock$finalize()}}
\item \href{#method-FileBlock-from_list}{\code{FileBlock$from_list()}}
\item \href{#method-FileBlock-to_list}{\code{FileBlock$to_list()}}
\item \href{#method-FileBlock-clone}{\code{FileBlock$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="teal.reporter" data-topic="ContentBlock" data-id="get_content"><a href='../../teal.reporter/html/ContentBlock.html#method-ContentBlock-get_content'><code>teal.reporter::ContentBlock$get_content()</code></a></li>
<li><span class="pkg-link" data-pkg="teal.reporter" data-topic="ContentBlock" data-id="initialize"><a href='../../teal.reporter/html/ContentBlock.html#method-ContentBlock-initialize'><code>teal.reporter::ContentBlock$initialize()</code></a></li>
<li><span class="pkg-link" data-pkg="teal.reporter" data-topic="ContentBlock" data-id="set_content"><a href='../../teal.reporter/html/ContentBlock.html#method-ContentBlock-set_content'><code>teal.reporter::ContentBlock$set_content()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileBlock-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-FileBlock-finalize}{}}}
\subsection{Method \code{finalize()}}{
finalize of this \code{FileBlock}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileBlock$finalize()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Removes the temporary file created in the constructor.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileBlock-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-FileBlock-from_list}{}}}
\subsection{Method \code{from_list()}}{
Create the \code{FileBlock} from a list.
The list should contain one named field, \code{"path"}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileBlock$from_list(x, base_path = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\verb{named list} with one field \code{"path"}, a path to the file.}

\item{\code{base_path}}{\code{character} with a path to the dir with the file.
If the argument is \code{NULL} then only \code{x} is used. Default \code{NULL}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisibly self
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{block <- teal.reporter:::FileBlock$new()
file_path <- tempfile(fileext = ".png")
saveRDS(iris, file_path)
block$from_list(list(path = file_path))
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileBlock-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-FileBlock-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert the \code{FileBlock} to a list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileBlock$to_list(base_path = ".")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{base_path}}{\code{character} with a path to the dir with the file.
If the argument is \code{NULL} then only basename is returned. Default \code{.}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\verb{named list} with a path to the file.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{block <- teal.reporter:::FileBlock$new()
block$to_list()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileBlock-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FileBlock-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileBlock$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
