% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTMLBlock.R
\docType{class}
\name{HTMLBlock}
\alias{HTMLBlock}
\title{\code{HTMLBlock}}
\description{
Specialized \code{FileBlock} for managing HTML content in reports.
It's designed to handle various HTML content, and render the report as HTML,
however \code{htmlwidgets} objects can also be rendered to static document-ready format.
}
\examples{

## ------------------------------------------------
## Method `HTMLBlock$set_content`
## ------------------------------------------------

HTMLBlock <- getFromNamespace("HTMLBlock", "teal.reporter")
block <- HTMLBlock$new()
block$set_content(shiny::div("HTML Content"))

}
\keyword{internal}
\section{Super classes}{
\code{\link[teal.reporter:ContentBlock]{teal.reporter::ContentBlock}} -> \code{\link[teal.reporter:FileBlock]{teal.reporter::FileBlock}} -> \code{HTMLBlock}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-HTMLBlock-new}{\code{HTMLBlock$new()}}
\item \href{#method-HTMLBlock-set_content}{\code{HTMLBlock$set_content()}}
\item \href{#method-HTMLBlock-clone}{\code{HTMLBlock$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="teal.reporter" data-topic="ContentBlock" data-id="get_content"><a href='../../teal.reporter/html/ContentBlock.html#method-ContentBlock-get_content'><code>teal.reporter::ContentBlock$get_content()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.reporter" data-topic="FileBlock" data-id="finalize"><a href='../../teal.reporter/html/FileBlock.html#method-FileBlock-finalize'><code>teal.reporter::FileBlock$finalize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.reporter" data-topic="FileBlock" data-id="from_list"><a href='../../teal.reporter/html/FileBlock.html#method-FileBlock-from_list'><code>teal.reporter::FileBlock$from_list()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.reporter" data-topic="FileBlock" data-id="to_list"><a href='../../teal.reporter/html/FileBlock.html#method-FileBlock-to_list'><code>teal.reporter::FileBlock$to_list()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HTMLBlock-new"></a>}}
\if{latex}{\out{\hypertarget{method-HTMLBlock-new}{}}}
\subsection{Method \code{new()}}{
Initialize a \code{HTMLBlock} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HTMLBlock$new(content)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{content}}{An object that can be rendered as a HTML content assigned to
this \code{HTMLBlock}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Object of class \code{HTMLBlock}, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HTMLBlock-set_content"></a>}}
\if{latex}{\out{\hypertarget{method-HTMLBlock-set_content}{}}}
\subsection{Method \code{set_content()}}{
Sets content of this \code{HTMLBlock}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HTMLBlock$set_content(content)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{content}}{An object that can be rendered as a HTML content
assigned to this \code{HTMLBlock}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{self}, invisibly.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{HTMLBlock <- getFromNamespace("HTMLBlock", "teal.reporter")
block <- HTMLBlock$new()
block$set_content(shiny::div("HTML Content"))

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HTMLBlock-clone"></a>}}
\if{latex}{\out{\hypertarget{method-HTMLBlock-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HTMLBlock$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
