% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Archiver.R
\name{JSONArchiver}
\alias{JSONArchiver}
\title{\code{JSONArchiver}}
\description{
\code{JSONArchiver}

\code{JSONArchiver}
}
\examples{

## ------------------------------------------------
## Method `JSONArchiver$new`
## ------------------------------------------------

archiver <- teal.reporter:::JSONArchiver$new()


## ------------------------------------------------
## Method `JSONArchiver$write`
## ------------------------------------------------

card1 <- teal.reporter:::ReportCard$new()

card1$append_text("Header 2 text", "header2")
card1$append_text("A paragraph of default text", "header2")
card1$append_plot(
 ggplot2::ggplot(iris, ggplot2::aes(x = Petal.Length)) + ggplot2::geom_histogram()
)

card2 <- teal.reporter:::ReportCard$new()

card2$append_text("Header 2 text", "header2")
card2$append_text("A paragraph of default text", "header2")
lyt <- rtables::analyze(rtables::split_rows_by(rtables::basic_table(), "Day"), "Ozone", afun = mean)
table_res2 <- rtables::build_table(lyt, airquality)
card2$append_table(table_res2)
card2$append_table(iris)

reporter <- teal.reporter:::Reporter$new()
reporter$append_cards(list(card1, card2))

archiver <- JSONArchiver$new()
archiver$write(reporter)

archiver$read()$get_cards()[[1]]$get_content()
archiver$get_output_dir()
zip::zipr("../test.zip", list.files(archiver$get_output_dir(), full.names = TRUE))
archiver$read("../test.zip")$get_cards()[[1]]$get_content()[[3]]$get_content()

## ------------------------------------------------
## Method `JSONArchiver$read`
## ------------------------------------------------

card1 <- teal.reporter:::ReportCard$new()

card1$append_text("Header 2 text", "header2")
card1$append_text("A paragraph of default text", "header2")
card1$append_plot(
 ggplot2::ggplot(iris, ggplot2::aes(x = Petal.Length)) + ggplot2::geom_histogram()
)

card2 <- teal.reporter:::ReportCard$new()

card2$append_text("Header 2 text", "header2")
card2$append_text("A paragraph of default text", "header2")
lyt <- rtables::analyze(rtables::split_rows_by(rtables::basic_table(), "Day"), "Ozone", afun = mean)
table_res2 <- rtables::build_table(lyt, airquality)
card2$append_table(table_res2)
card2$append_table(iris)

reporter <- teal.reporter:::Reporter$new()
reporter$append_cards(list(card1, card2))

archiver <- JSONArchiver$new()
archiver$write(reporter)

archiver$read()$get_cards()[[1]]$get_content()
archiver$get_output_dir()
zip::zipr("../test.zip", list.files(archiver$get_output_dir(), full.names = TRUE))
archiver$read("../test.zip")$get_cards()[[1]]$get_content()[[3]]$get_content()

## ------------------------------------------------
## Method `JSONArchiver$get_output_dir`
## ------------------------------------------------

archiver <- teal.reporter:::Archiver$new()
archiver$get_output_dir()
}
\keyword{internal}
\section{Super class}{
\code{\link[teal.reporter:Archiver]{teal.reporter::Archiver}} -> \code{JSONArchiver}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-JSONArchiver-new}{\code{JSONArchiver$new()}}
\item \href{#method-JSONArchiver-finalize}{\code{JSONArchiver$finalize()}}
\item \href{#method-JSONArchiver-write}{\code{JSONArchiver$write()}}
\item \href{#method-JSONArchiver-read}{\code{JSONArchiver$read()}}
\item \href{#method-JSONArchiver-get_output_dir}{\code{JSONArchiver$get_output_dir()}}
\item \href{#method-JSONArchiver-clone}{\code{JSONArchiver$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JSONArchiver-new"></a>}}
\if{latex}{\out{\hypertarget{method-JSONArchiver-new}{}}}
\subsection{Method \code{new()}}{
Returns a \code{JSONArchiver} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JSONArchiver$new(version)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a \code{JSONArchiver} object
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{archiver <- teal.reporter:::JSONArchiver$new()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JSONArchiver-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-JSONArchiver-finalize}{}}}
\subsection{Method \code{finalize()}}{
Finalizes a \code{Archiver} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JSONArchiver$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JSONArchiver-write"></a>}}
\if{latex}{\out{\hypertarget{method-JSONArchiver-write}{}}}
\subsection{Method \code{write()}}{
write a \code{Reporter} instance in to this \code{JSONArchiver} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JSONArchiver$write(reporter, report_params = list(), datasets = NULL)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisibly self
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{card1 <- teal.reporter:::ReportCard$new()

card1$append_text("Header 2 text", "header2")
card1$append_text("A paragraph of default text", "header2")
card1$append_plot(
 ggplot2::ggplot(iris, ggplot2::aes(x = Petal.Length)) + ggplot2::geom_histogram()
)

card2 <- teal.reporter:::ReportCard$new()

card2$append_text("Header 2 text", "header2")
card2$append_text("A paragraph of default text", "header2")
lyt <- rtables::analyze(rtables::split_rows_by(rtables::basic_table(), "Day"), "Ozone", afun = mean)
table_res2 <- rtables::build_table(lyt, airquality)
card2$append_table(table_res2)
card2$append_table(iris)

reporter <- teal.reporter:::Reporter$new()
reporter$append_cards(list(card1, card2))

archiver <- JSONArchiver$new()
archiver$write(reporter)

archiver$read()$get_cards()[[1]]$get_content()
archiver$get_output_dir()
zip::zipr("../test.zip", list.files(archiver$get_output_dir(), full.names = TRUE))
archiver$read("../test.zip")$get_cards()[[1]]$get_content()[[3]]$get_content()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JSONArchiver-read"></a>}}
\if{latex}{\out{\hypertarget{method-JSONArchiver-read}{}}}
\subsection{Method \code{read()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JSONArchiver$read(path2zip = NULL)}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{card1 <- teal.reporter:::ReportCard$new()

card1$append_text("Header 2 text", "header2")
card1$append_text("A paragraph of default text", "header2")
card1$append_plot(
 ggplot2::ggplot(iris, ggplot2::aes(x = Petal.Length)) + ggplot2::geom_histogram()
)

card2 <- teal.reporter:::ReportCard$new()

card2$append_text("Header 2 text", "header2")
card2$append_text("A paragraph of default text", "header2")
lyt <- rtables::analyze(rtables::split_rows_by(rtables::basic_table(), "Day"), "Ozone", afun = mean)
table_res2 <- rtables::build_table(lyt, airquality)
card2$append_table(table_res2)
card2$append_table(iris)

reporter <- teal.reporter:::Reporter$new()
reporter$append_cards(list(card1, card2))

archiver <- JSONArchiver$new()
archiver$write(reporter)

archiver$read()$get_cards()[[1]]$get_content()
archiver$get_output_dir()
zip::zipr("../test.zip", list.files(archiver$get_output_dir(), full.names = TRUE))
archiver$read("../test.zip")$get_cards()[[1]]$get_content()[[3]]$get_content()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JSONArchiver-get_output_dir"></a>}}
\if{latex}{\out{\hypertarget{method-JSONArchiver-get_output_dir}{}}}
\subsection{Method \code{get_output_dir()}}{
get \code{output_dir} field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JSONArchiver$get_output_dir()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} a \code{output_dir} field path.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{archiver <- teal.reporter:::Archiver$new()
archiver$get_output_dir()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JSONArchiver-clone"></a>}}
\if{latex}{\out{\hypertarget{method-JSONArchiver-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JSONArchiver$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
