% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Archiver.R
\name{JSONArchiver}
\alias{JSONArchiver}
\title{\code{JSONArchiver}}
\description{
\code{JSONArchiver}

\code{JSONArchiver}
}
\examples{

## ------------------------------------------------
## Method `JSONArchiver$write`
## ------------------------------------------------

card1 <- teal.reporter:::ReportCard$new()

card1$append_text("Header 2 text", "header2")
card1$append_text("A paragraph of default text", "header2")
card1$append_plot(
 ggplot2::ggplot(iris, ggplot2::aes(x = Petal.Length)) + ggplot2::geom_histogram()
)

card2 <- teal.reporter:::ReportCard$new()

card2$append_text("Header 2 text", "header2")
card2$append_text("A paragraph of default text", "header2")
lyt <- rtables::analyze(rtables::split_rows_by(rtables::basic_table(), "Day"), "Ozone", afun = mean)
table_res2 <- rtables::build_table(lyt, airquality)
card2$append_table(table_res2)
card2$append_table(iris)

reporter <- teal.reporter:::Reporter$new()
reporter$append_cards(list(card1, card2))

archiver <- JSONArchiver$new()
archiver$write(reporter)

archiver$read()$get_cards()[[1]]$get_content()
archiver$get_output_dir()
zip::zipr("../test.zip", list.files(archiver$get_output_dir(), full.names = TRUE))
archiver$read("../test.zip")$get_cards()[[1]]$get_content()[[3]]$get_content()

## ------------------------------------------------
## Method `JSONArchiver$read`
## ------------------------------------------------

card1 <- teal.reporter:::ReportCard$new()

card1$append_text("Header 2 text", "header2")
card1$append_text("A paragraph of default text", "header2")
card1$append_plot(
 ggplot2::ggplot(iris, ggplot2::aes(x = Petal.Length)) + ggplot2::geom_histogram()
)

card2 <- teal.reporter:::ReportCard$new()

card2$append_text("Header 2 text", "header2")
card2$append_text("A paragraph of default text", "header2")
lyt <- rtables::analyze(rtables::split_rows_by(rtables::basic_table(), "Day"), "Ozone", afun = mean)
table_res2 <- rtables::build_table(lyt, airquality)
card2$append_table(table_res2)
card2$append_table(iris)

reporter <- teal.reporter:::Reporter$new()
reporter$append_cards(list(card1, card2))

archiver <- JSONArchiver$new()
archiver$write(reporter)

archiver$read()$get_cards()[[1]]$get_content()
archiver$get_output_dir()
zip::zipr("../test.zip", list.files(archiver$get_output_dir(), full.names = TRUE))
archiver$read("../test.zip")$get_cards()[[1]]$get_content()[[3]]$get_content()
}
\keyword{internal}
\section{Super classes}{
\code{\link[teal.reporter:Archiver]{teal.reporter::Archiver}} -> \code{\link[teal.reporter:RDSArchiver]{teal.reporter::RDSArchiver}} -> \code{JSONArchiver}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-JSONArchiver-write}{\code{JSONArchiver$write()}}
\item \href{#method-JSONArchiver-read}{\code{JSONArchiver$read()}}
\item \href{#method-JSONArchiver-clone}{\code{JSONArchiver$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="teal.reporter" data-topic="RDSArchiver" data-id="finalize"><a href='../../teal.reporter/html/RDSArchiver.html#method-RDSArchiver-finalize'><code>teal.reporter::RDSArchiver$finalize()</code></a></li>
<li><span class="pkg-link" data-pkg="teal.reporter" data-topic="RDSArchiver" data-id="get_output_dir"><a href='../../teal.reporter/html/RDSArchiver.html#method-RDSArchiver-get_output_dir'><code>teal.reporter::RDSArchiver$get_output_dir()</code></a></li>
<li><span class="pkg-link" data-pkg="teal.reporter" data-topic="RDSArchiver" data-id="initialize"><a href='../../teal.reporter/html/RDSArchiver.html#method-RDSArchiver-initialize'><code>teal.reporter::RDSArchiver$initialize()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JSONArchiver-write"></a>}}
\if{latex}{\out{\hypertarget{method-JSONArchiver-write}{}}}
\subsection{Method \code{write()}}{
write a \code{Reporter} instance in to this \code{JSONArchiver} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JSONArchiver$write(reporter)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reporter}}{\code{Reporter} instance.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisibly self
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{card1 <- teal.reporter:::ReportCard$new()

card1$append_text("Header 2 text", "header2")
card1$append_text("A paragraph of default text", "header2")
card1$append_plot(
 ggplot2::ggplot(iris, ggplot2::aes(x = Petal.Length)) + ggplot2::geom_histogram()
)

card2 <- teal.reporter:::ReportCard$new()

card2$append_text("Header 2 text", "header2")
card2$append_text("A paragraph of default text", "header2")
lyt <- rtables::analyze(rtables::split_rows_by(rtables::basic_table(), "Day"), "Ozone", afun = mean)
table_res2 <- rtables::build_table(lyt, airquality)
card2$append_table(table_res2)
card2$append_table(iris)

reporter <- teal.reporter:::Reporter$new()
reporter$append_cards(list(card1, card2))

archiver <- JSONArchiver$new()
archiver$write(reporter)

archiver$read()$get_cards()[[1]]$get_content()
archiver$get_output_dir()
zip::zipr("../test.zip", list.files(archiver$get_output_dir(), full.names = TRUE))
archiver$read("../test.zip")$get_cards()[[1]]$get_content()[[3]]$get_content()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JSONArchiver-read"></a>}}
\if{latex}{\out{\hypertarget{method-JSONArchiver-read}{}}}
\subsection{Method \code{read()}}{
read a \code{Reporter} instance from a directory with \code{JSONArchiver}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JSONArchiver$read(path2zip = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path2zip}}{\code{character(1)} a path to the directory with all proper files.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{Reporter} instance.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{card1 <- teal.reporter:::ReportCard$new()

card1$append_text("Header 2 text", "header2")
card1$append_text("A paragraph of default text", "header2")
card1$append_plot(
 ggplot2::ggplot(iris, ggplot2::aes(x = Petal.Length)) + ggplot2::geom_histogram()
)

card2 <- teal.reporter:::ReportCard$new()

card2$append_text("Header 2 text", "header2")
card2$append_text("A paragraph of default text", "header2")
lyt <- rtables::analyze(rtables::split_rows_by(rtables::basic_table(), "Day"), "Ozone", afun = mean)
table_res2 <- rtables::build_table(lyt, airquality)
card2$append_table(table_res2)
card2$append_table(iris)

reporter <- teal.reporter:::Reporter$new()
reporter$append_cards(list(card1, card2))

archiver <- JSONArchiver$new()
archiver$write(reporter)

archiver$read()$get_cards()[[1]]$get_content()
archiver$get_output_dir()
zip::zipr("../test.zip", list.files(archiver$get_output_dir(), full.names = TRUE))
archiver$read("../test.zip")$get_cards()[[1]]$get_content()[[3]]$get_content()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JSONArchiver-clone"></a>}}
\if{latex}{\out{\hypertarget{method-JSONArchiver-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JSONArchiver$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
