% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportDocument.R
\name{keep_in_report}
\alias{keep_in_report}
\title{Keep Objects In Report}
\usage{
keep_in_report(object, keep = TRUE)
}
\arguments{
\item{object}{An R object, typically an element intended for a \code{doc}.}

\item{keep}{(\code{logical}) If \code{TRUE} (default), the object is marked to be kept in the \code{.Rmd} source;
if \code{FALSE}, it's marked for printing only in the output document (and not in the \code{.Rmd} source,
though its print output will be in the rendered document).}
}
\value{
The input \code{object} with its "keep" attribute modified.
}
\description{
Utility function to change behavior of \code{doc} elements to be
kept (\code{keep = TRUE}) or discarded (\code{keep = FALSE}) from the final \code{.Rmd} file containing the downloaded report.
}
\details{
By default, R objects like \code{summary} outputs are only printed in the output document but their
code is not included in the \code{.Rmd} report source. Text elements (character strings) and \code{code_chunk}
objects are, by default, kept both in the output document and the \code{.Rmd} report source.
This function allows overriding the default behavior for specific objects.
}
\examples{
item <- summary(iris)
item <- keep_in_report(item, TRUE)
attributes(item)$keep

}
