% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewpageBlock.R
\name{NewpageBlock}
\alias{NewpageBlock}
\title{\code{NewpageBlock}}
\description{
\code{NewpageBlock}

\code{NewpageBlock}
}
\examples{

## ------------------------------------------------
## Method `NewpageBlock$new`
## ------------------------------------------------

NewpageBlock <- getFromNamespace("NewpageBlock", "teal.reporter")
block <- NewpageBlock$new()

}
\keyword{internal}
\section{Super class}{
\code{\link[teal.reporter:ContentBlock]{teal.reporter::ContentBlock}} -> \code{NewpageBlock}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NewpageBlock-new}{\code{NewpageBlock$new()}}
\item \href{#method-NewpageBlock-clone}{\code{NewpageBlock$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="teal.reporter" data-topic="ContentBlock" data-id="from_list"><a href='../../teal.reporter/html/ContentBlock.html#method-ContentBlock-from_list'><code>teal.reporter::ContentBlock$from_list()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.reporter" data-topic="ContentBlock" data-id="get_content"><a href='../../teal.reporter/html/ContentBlock.html#method-ContentBlock-get_content'><code>teal.reporter::ContentBlock$get_content()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.reporter" data-topic="ContentBlock" data-id="set_content"><a href='../../teal.reporter/html/ContentBlock.html#method-ContentBlock-set_content'><code>teal.reporter::ContentBlock$set_content()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.reporter" data-topic="ContentBlock" data-id="to_list"><a href='../../teal.reporter/html/ContentBlock.html#method-ContentBlock-to_list'><code>teal.reporter::ContentBlock$to_list()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NewpageBlock-new"></a>}}
\if{latex}{\out{\hypertarget{method-NewpageBlock-new}{}}}
\subsection{Method \code{new()}}{
Returns a \code{NewpageBlock} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NewpageBlock$new()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Returns a \code{NewpageBlock} object with no content and the default style.
}

\subsection{Returns}{
\code{NewpageBlock}.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{NewpageBlock <- getFromNamespace("NewpageBlock", "teal.reporter")
block <- NewpageBlock$new()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NewpageBlock-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NewpageBlock-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NewpageBlock$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
