% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewpageBlock.R
\name{NewpageBlock}
\alias{NewpageBlock}
\title{\code{NewpageBlock}}
\description{
\code{NewpageBlock}

\code{NewpageBlock}
}
\examples{

## ------------------------------------------------
## Method `NewpageBlock$new`
## ------------------------------------------------

block <- teal.reporter:::NewpageBlock$new()


## ------------------------------------------------
## Method `NewpageBlock$get_content`
## ------------------------------------------------

block <- teal.reporter:::NewpageBlock$new()
block$get_content()

}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NewpageBlock$new()}}
\item \href{#method-get_content}{\code{NewpageBlock$get_content()}}
\item \href{#method-clone}{\code{NewpageBlock$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Returns a \code{NewpageBlock} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NewpageBlock$new()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Returns a \code{NewpageBlock} object with no content and the default style.
}

\subsection{Returns}{
\code{NewpageBlock}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{block <- teal.reporter:::NewpageBlock$new()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_content"></a>}}
\if{latex}{\out{\hypertarget{method-get_content}{}}}
\subsection{Method \code{get_content()}}{
Returns the content of this \code{NewpageBlock}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NewpageBlock$get_content()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the content of this \code{NewpageBlock}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{block <- teal.reporter:::NewpageBlock$new()
block$get_content()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NewpageBlock$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
