% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{panel_item}
\alias{panel_item}
\title{Panel group widget}
\usage{
panel_item(title, ..., collapsed = TRUE, input_id = NULL)
}
\arguments{
\item{title}{(\code{character}) title of panel}

\item{...}{content of panel}

\item{collapsed}{(\code{logical}, optional)
whether to initially collapse panel}

\item{input_id}{(\code{character}, optional)
name of the panel item element. If supplied, this will register a shiny input variable that
indicates whether the panel item is open or collapsed and is accessed with \code{input$input_id}.}
}
\value{
\code{shiny.tag}.
}
\description{
Panel group widget
}
\keyword{internal}
