% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PictureBlock.R
\name{PictureBlock}
\alias{PictureBlock}
\title{\code{PictureBlock}}
\description{
\code{PictureBlock}

\code{PictureBlock}
}
\examples{

## ------------------------------------------------
## Method `PictureBlock$new`
## ------------------------------------------------

block <- teal.reporter:::PictureBlock$new()


## ------------------------------------------------
## Method `PictureBlock$set_content`
## ------------------------------------------------

block <- teal.reporter:::PictureBlock$new()
block$set_content("Base64 encoded picture")


## ------------------------------------------------
## Method `PictureBlock$get_content`
## ------------------------------------------------

block <- teal.reporter:::PictureBlock$new()
block$get_content()

}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PictureBlock$new()}}
\item \href{#method-set_content}{\code{PictureBlock$set_content()}}
\item \href{#method-get_content}{\code{PictureBlock$get_content()}}
\item \href{#method-clone}{\code{PictureBlock$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Returns a \code{PictureBlock} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PictureBlock$new()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Returns a \code{PictureBlock} object with no content and the default style.
}

\subsection{Returns}{
\code{PictureBlock}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{block <- teal.reporter:::PictureBlock$new()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_content"></a>}}
\if{latex}{\out{\hypertarget{method-set_content}{}}}
\subsection{Method \code{set_content()}}{
Sets the content of this \code{PictureBlock}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PictureBlock$set_content(content)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{content}}{(\code{character(1)}) a string assigned to this \code{PictureBlock}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
throws if argument is not \code{character(1)}.
}

\subsection{Returns}{
invisibly self
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{block <- teal.reporter:::PictureBlock$new()
block$set_content("Base64 encoded picture")

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_content"></a>}}
\if{latex}{\out{\hypertarget{method-get_content}{}}}
\subsection{Method \code{get_content()}}{
Returns the content of this \code{PictureBlock}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PictureBlock$get_content()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the content of this \code{PictureBlock}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{block <- teal.reporter:::PictureBlock$new()
block$get_content()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PictureBlock$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
