% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PictureBlock.R
\name{PictureBlock}
\alias{PictureBlock}
\title{\code{PictureBlock}}
\description{
\code{PictureBlock}

\code{PictureBlock}
}
\examples{

## ------------------------------------------------
## Method `PictureBlock$set_content`
## ------------------------------------------------

block <- teal.reporter:::PictureBlock$new()
block$set_content(ggplot2::ggplot(iris))

block <- teal.reporter:::PictureBlock$new()
block$set_content(lattice::bwplot(1))

block <- teal.reporter:::PictureBlock$new()
block$set_content(ggplot2::ggplotGrob(ggplot2::ggplot(iris)))

## ------------------------------------------------
## Method `PictureBlock$set_title`
## ------------------------------------------------

block <- teal.reporter:::PictureBlock$new()
block$set_title("Title")


## ------------------------------------------------
## Method `PictureBlock$get_title`
## ------------------------------------------------

block <- teal.reporter:::PictureBlock$new()
block$get_title()


## ------------------------------------------------
## Method `PictureBlock$set_dim`
## ------------------------------------------------

block <- teal.reporter:::PictureBlock$new()
block$set_dim(c(800, 600))


## ------------------------------------------------
## Method `PictureBlock$get_dim`
## ------------------------------------------------

block <- teal.reporter:::PictureBlock$new()
block$get_dim()

}
\keyword{internal}
\section{Super classes}{
\code{\link[teal.reporter:ContentBlock]{teal.reporter::ContentBlock}} -> \code{\link[teal.reporter:FileBlock]{teal.reporter::FileBlock}} -> \code{PictureBlock}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PictureBlock-new}{\code{PictureBlock$new()}}
\item \href{#method-PictureBlock-set_content}{\code{PictureBlock$set_content()}}
\item \href{#method-PictureBlock-set_title}{\code{PictureBlock$set_title()}}
\item \href{#method-PictureBlock-get_title}{\code{PictureBlock$get_title()}}
\item \href{#method-PictureBlock-set_dim}{\code{PictureBlock$set_dim()}}
\item \href{#method-PictureBlock-get_dim}{\code{PictureBlock$get_dim()}}
\item \href{#method-PictureBlock-clone}{\code{PictureBlock$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="teal.reporter" data-topic="ContentBlock" data-id="get_content"><a href='../../teal.reporter/html/ContentBlock.html#method-ContentBlock-get_content'><code>teal.reporter::ContentBlock$get_content()</code></a></li>
<li><span class="pkg-link" data-pkg="teal.reporter" data-topic="FileBlock" data-id="finalize"><a href='../../teal.reporter/html/FileBlock.html#method-FileBlock-finalize'><code>teal.reporter::FileBlock$finalize()</code></a></li>
<li><span class="pkg-link" data-pkg="teal.reporter" data-topic="FileBlock" data-id="from_list"><a href='../../teal.reporter/html/FileBlock.html#method-FileBlock-from_list'><code>teal.reporter::FileBlock$from_list()</code></a></li>
<li><span class="pkg-link" data-pkg="teal.reporter" data-topic="FileBlock" data-id="to_list"><a href='../../teal.reporter/html/FileBlock.html#method-FileBlock-to_list'><code>teal.reporter::FileBlock$to_list()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PictureBlock-new"></a>}}
\if{latex}{\out{\hypertarget{method-PictureBlock-new}{}}}
\subsection{Method \code{new()}}{
Returns a new \code{PictureBlock} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PictureBlock$new(plot)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plot}}{(\code{ggplot}, \code{grid}) a picture in this \code{PictureBlock}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a \code{PictureBlock} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PictureBlock-set_content"></a>}}
\if{latex}{\out{\hypertarget{method-PictureBlock-set_content}{}}}
\subsection{Method \code{set_content()}}{
Sets the content of this \code{PictureBlock}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PictureBlock$set_content(content)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{content}}{(\code{ggplot}, \code{grob}, \code{trellis}) a picture in this \code{PictureBlock}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
throws if argument is not a \code{ggplot}, \code{grob} or \code{trellis} plot.
}

\subsection{Returns}{
invisibly self
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{block <- teal.reporter:::PictureBlock$new()
block$set_content(ggplot2::ggplot(iris))

block <- teal.reporter:::PictureBlock$new()
block$set_content(lattice::bwplot(1))

block <- teal.reporter:::PictureBlock$new()
block$set_content(ggplot2::ggplotGrob(ggplot2::ggplot(iris)))
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PictureBlock-set_title"></a>}}
\if{latex}{\out{\hypertarget{method-PictureBlock-set_title}{}}}
\subsection{Method \code{set_title()}}{
Sets the title of this \code{PictureBlock}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PictureBlock$set_title(title)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{(\code{character(1)}) a string assigned to this \code{PictureBlock}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
throws if argument is not \code{character(1)}.
}

\subsection{Returns}{
invisibly self
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{block <- teal.reporter:::PictureBlock$new()
block$set_title("Title")

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PictureBlock-get_title"></a>}}
\if{latex}{\out{\hypertarget{method-PictureBlock-get_title}{}}}
\subsection{Method \code{get_title()}}{
Returns the title of this \code{PictureBlock}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PictureBlock$get_title()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the content of this \code{PictureBlock}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{block <- teal.reporter:::PictureBlock$new()
block$get_title()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PictureBlock-set_dim"></a>}}
\if{latex}{\out{\hypertarget{method-PictureBlock-set_dim}{}}}
\subsection{Method \code{set_dim()}}{
Sets the dimensions of this \code{PictureBlock}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PictureBlock$set_dim(dim)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dim}}{\code{numeric} figure dimensions (width and height) in pixels, length 2.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{self}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{block <- teal.reporter:::PictureBlock$new()
block$set_dim(c(800, 600))

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PictureBlock-get_dim"></a>}}
\if{latex}{\out{\hypertarget{method-PictureBlock-get_dim}{}}}
\subsection{Method \code{get_dim()}}{
Returns the dimensions of this \code{PictureBlock}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PictureBlock$get_dim()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{numeric} the array of 2 numeric values representing width and height in pixels.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{block <- teal.reporter:::PictureBlock$new()
block$get_dim()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PictureBlock-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PictureBlock-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PictureBlock$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
