% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_utils.R
\name{print.rmd_yaml_header}
\alias{print.rmd_yaml_header}
\title{Print method for the \code{yaml_header} class}
\usage{
\method{print}{rmd_yaml_header}(x, ...)
}
\arguments{
\item{x}{\code{rmd_yaml_header} class object.}

\item{...}{optional text.}
}
\value{
\code{NULL}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
input <- list(author = "", output = "pdf_document", toc = TRUE, keep_tex = TRUE)
out <- as_yaml_auto(input)
out
print(out)
}
