% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.r
\name{render}
\alias{render}
\title{Render \code{teal_card}}
\usage{
render(
  input,
  output_dir = NULL,
  global_knitr = getOption("teal.reporter.global_knitr"),
  ...
)
}
\arguments{
\item{input}{(\code{teal_report}, \code{teal_code}) object to render}

\item{output_dir}{The output directory for the rendered \code{output_file}.
This allows for a choice of an alternate directory to which the output file
should be written (the default output directory of that of the input file).
If a path is provided with a filename in \code{output_file} the directory
specified here will take precedence. Please note that any directory path
provided will create any necessary directories if they do not exist.}

\item{global_knitr}{(\code{list}) # todo}

\item{...}{arguments passed to \code{rmarkdown::render}.}
}
\description{
Render \code{teal_card}
}
