% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Renderer.R
\name{Renderer}
\alias{Renderer}
\title{\code{Renderer}}
\description{
\code{Renderer}

\code{Renderer}
}
\examples{

## ------------------------------------------------
## Method `Renderer$new`
## ------------------------------------------------

renderer <- teal.reporter:::Renderer$new()

## ------------------------------------------------
## Method `Renderer$renderRmd`
## ------------------------------------------------

card1 <- teal.reporter:::ReportCard$new()

card1$append_text("Header 2 text", "header2")
card1$append_text("A paragraph of default text")
card1$append_plot(
 ggplot2::ggplot(iris, ggplot2::aes(x = Petal.Length)) + ggplot2::geom_histogram()
)

card2 <- teal.reporter:::ReportCard$new()

card2$append_text("Header 2 text", "header2")
card2$append_text("A paragraph of default text", "header2")
lyt <- rtables::analyze(rtables::split_rows_by(rtables::basic_table(), "Day"), "Ozone", afun = mean)
table_res2 <- rtables::build_table(lyt, airquality)
card2$append_table(table_res2)
card2$append_table(iris)

reporter <- teal.reporter:::Reporter$new()
reporter$append_cards(list(card1, card2))

yaml_l <- list(
  author = teal.reporter:::yaml_quoted("NEST"),
  title = teal.reporter:::yaml_quoted("Report"),
  date = teal.reporter:::yaml_quoted("07/04/2019"),
  output = list(pdf_document = list(keep_tex = TRUE))
)

yaml_header <- teal.reporter:::md_header(yaml::as.yaml(yaml_l))
result_path <- teal.reporter:::Renderer$new()$renderRmd(reporter$get_blocks(), yaml_header)

## ------------------------------------------------
## Method `Renderer$render`
## ------------------------------------------------

card1 <- teal.reporter:::ReportCard$new()
card1$append_text("Header 2 text", "header2")
card1$append_text("A paragraph of default text")
card1$append_plot(
 ggplot2::ggplot(iris, ggplot2::aes(x = Petal.Length)) + ggplot2::geom_histogram()
)

card2 <- teal.reporter:::ReportCard$new()
card2$append_text("Header 2 text", "header2")
card2$append_text("A paragraph of default text", "header2")
lyt <- rtables::analyze(rtables::split_rows_by(rtables::basic_table(), "Day"), "Ozone", afun = mean)
table_res2 <- rtables::build_table(lyt, airquality)
card2$append_table(table_res2)
card2$append_table(iris)

reporter <- teal.reporter:::Reporter$new()
reporter$append_cards(list(card1, card2))

yaml_l <- list(
  author = teal.reporter:::yaml_quoted("NEST"),
  title = teal.reporter:::yaml_quoted("Report"),
  date = teal.reporter:::yaml_quoted("07/04/2019"),
  output = list(pdf_document = list(keep_tex = TRUE))
)

yaml_header <- teal.reporter:::md_header(yaml::as.yaml(yaml_l))
result_path <- teal.reporter:::Renderer$new()$render(reporter$get_blocks(), yaml_header)

## ------------------------------------------------
## Method `Renderer$get_output_dir`
## ------------------------------------------------

renderer <- teal.reporter:::Renderer$new()
renderer$get_output_dir()
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Renderer$new()}}
\item \href{#method-finalize}{\code{Renderer$finalize()}}
\item \href{#method-renderRmd}{\code{Renderer$renderRmd()}}
\item \href{#method-render}{\code{Renderer$render()}}
\item \href{#method-get_output_dir}{\code{Renderer$get_output_dir()}}
\item \href{#method-clone}{\code{Renderer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Returns a \code{Renderer} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Renderer$new()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Returns a \code{Renderer} object.
}

\subsection{Returns}{
\code{Renderer} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{renderer <- teal.reporter:::Renderer$new()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
Finalizes a \code{Renderer} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Renderer$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-renderRmd"></a>}}
\if{latex}{\out{\hypertarget{method-renderRmd}{}}}
\subsection{Method \code{renderRmd()}}{
getting the \code{Rmd} text which could be easily rendered later.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Renderer$renderRmd(blocks, yaml_header)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{blocks}}{\code{list} of \code{c("TextBlock", "PictureBlock", "NewpageBlock")} objects.}

\item{\code{yaml_header}}{\code{character} a \code{rmarkdown} \code{yaml} header.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{character} a \code{Rmd} text (\code{yaml} header + body), ready to be rendered.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{card1 <- teal.reporter:::ReportCard$new()

card1$append_text("Header 2 text", "header2")
card1$append_text("A paragraph of default text")
card1$append_plot(
 ggplot2::ggplot(iris, ggplot2::aes(x = Petal.Length)) + ggplot2::geom_histogram()
)

card2 <- teal.reporter:::ReportCard$new()

card2$append_text("Header 2 text", "header2")
card2$append_text("A paragraph of default text", "header2")
lyt <- rtables::analyze(rtables::split_rows_by(rtables::basic_table(), "Day"), "Ozone", afun = mean)
table_res2 <- rtables::build_table(lyt, airquality)
card2$append_table(table_res2)
card2$append_table(iris)

reporter <- teal.reporter:::Reporter$new()
reporter$append_cards(list(card1, card2))

yaml_l <- list(
  author = teal.reporter:::yaml_quoted("NEST"),
  title = teal.reporter:::yaml_quoted("Report"),
  date = teal.reporter:::yaml_quoted("07/04/2019"),
  output = list(pdf_document = list(keep_tex = TRUE))
)

yaml_header <- teal.reporter:::md_header(yaml::as.yaml(yaml_l))
result_path <- teal.reporter:::Renderer$new()$renderRmd(reporter$get_blocks(), yaml_header)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render"></a>}}
\if{latex}{\out{\hypertarget{method-render}{}}}
\subsection{Method \code{render()}}{
Renders the content of this \code{Report} to the output file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Renderer$render(blocks, yaml_header, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{blocks}}{\code{list} of \code{c("TextBlock", "PictureBlock", "NewpageBlock")} objects.}

\item{\code{yaml_header}}{\code{character} an \code{rmarkdown} \code{yaml} header.}

\item{\code{...}}{\code{rmarkdown::render} arguments, \code{input} and \code{output_dir} should not be updated.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{character} path to the output
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{card1 <- teal.reporter:::ReportCard$new()
card1$append_text("Header 2 text", "header2")
card1$append_text("A paragraph of default text")
card1$append_plot(
 ggplot2::ggplot(iris, ggplot2::aes(x = Petal.Length)) + ggplot2::geom_histogram()
)

card2 <- teal.reporter:::ReportCard$new()
card2$append_text("Header 2 text", "header2")
card2$append_text("A paragraph of default text", "header2")
lyt <- rtables::analyze(rtables::split_rows_by(rtables::basic_table(), "Day"), "Ozone", afun = mean)
table_res2 <- rtables::build_table(lyt, airquality)
card2$append_table(table_res2)
card2$append_table(iris)

reporter <- teal.reporter:::Reporter$new()
reporter$append_cards(list(card1, card2))

yaml_l <- list(
  author = teal.reporter:::yaml_quoted("NEST"),
  title = teal.reporter:::yaml_quoted("Report"),
  date = teal.reporter:::yaml_quoted("07/04/2019"),
  output = list(pdf_document = list(keep_tex = TRUE))
)

yaml_header <- teal.reporter:::md_header(yaml::as.yaml(yaml_l))
result_path <- teal.reporter:::Renderer$new()$render(reporter$get_blocks(), yaml_header)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_output_dir"></a>}}
\if{latex}{\out{\hypertarget{method-get_output_dir}{}}}
\subsection{Method \code{get_output_dir()}}{
get \code{output_dir} field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Renderer$get_output_dir()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} a \code{output_dir} field path.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{renderer <- teal.reporter:::Renderer$new()
renderer$get_output_dir()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Renderer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
