% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Renderer.R
\name{Renderer}
\alias{Renderer}
\title{\code{Renderer}}
\description{
\code{Renderer}

\code{Renderer}
}
\details{
To access the default values for the \code{global_knitr} parameter,
use \code{getOption("teal.reporter.global_knitr")}. These defaults include:
\itemize{
\item \code{echo = TRUE}
\item \code{tidy.opts = list(width.cutoff = 60)}
\item \code{tidy = TRUE}  if \code{formatR} package is installed, \code{FALSE} otherwise
}
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Renderer-new}{\code{Renderer$new()}}
\item \href{#method-Renderer-finalize}{\code{Renderer$finalize()}}
\item \href{#method-Renderer-renderRmd}{\code{Renderer$renderRmd()}}
\item \href{#method-Renderer-render}{\code{Renderer$render()}}
\item \href{#method-Renderer-get_output_dir}{\code{Renderer$get_output_dir()}}
\item \href{#method-Renderer-clone}{\code{Renderer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Renderer-new"></a>}}
\if{latex}{\out{\hypertarget{method-Renderer-new}{}}}
\subsection{Method \code{new()}}{
Returns a \code{Renderer} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Renderer$new()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Returns a \code{Renderer} object.
}

\subsection{Returns}{
\code{Renderer} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Renderer-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-Renderer-finalize}{}}}
\subsection{Method \code{finalize()}}{
Finalizes a \code{Renderer} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Renderer$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Renderer-renderRmd"></a>}}
\if{latex}{\out{\hypertarget{method-Renderer-renderRmd}{}}}
\subsection{Method \code{renderRmd()}}{
getting the \code{Rmd} text which could be easily rendered later.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Renderer$renderRmd(
  blocks,
  yaml_header,
  global_knitr = getOption("teal.reporter.global_knitr")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{blocks}}{\code{list} of \code{c("TextBlock", "PictureBlock", "NewpageBlock")} objects.}

\item{\code{yaml_header}}{\code{character} an \code{rmarkdown} \code{yaml} header.}

\item{\code{global_knitr}}{\code{list} a of \code{knitr} parameters (passed to \code{knitr::opts_chunk$set})
for customizing the rendering process.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{character} a \code{Rmd} text (\code{yaml} header + body), ready to be rendered.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Renderer-render"></a>}}
\if{latex}{\out{\hypertarget{method-Renderer-render}{}}}
\subsection{Method \code{render()}}{
Renders the content of this \code{Report} to the output file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Renderer$render(
  blocks,
  yaml_header,
  global_knitr = getOption("teal.reporter.global_knitr"),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{blocks}}{\code{list} of \code{c("TextBlock", "PictureBlock", "NewpageBlock")} objects.}

\item{\code{yaml_header}}{\code{character} an \code{rmarkdown} \code{yaml} header.}

\item{\code{global_knitr}}{\code{list} a of \code{knitr} parameters (passed to \code{knitr::opts_chunk$set})
for customizing the rendering process.}

\item{\code{...}}{\code{rmarkdown::render} arguments, \code{input} and \code{output_dir} should not be updated.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{character} path to the output
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Renderer-get_output_dir"></a>}}
\if{latex}{\out{\hypertarget{method-Renderer-get_output_dir}{}}}
\subsection{Method \code{get_output_dir()}}{
get \code{output_dir} field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Renderer$get_output_dir()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} a \code{output_dir} field path.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Renderer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Renderer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Renderer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
