% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_reportable-report.R
\name{report<-}
\alias{report<-}
\title{Replace a report in \code{teal_reportable}}
\usage{
report(x) <- value
}
\arguments{
\item{x}{(\code{teal_reportable})}

\item{value}{(\code{ReportDocument})}
}
\value{
\code{teal_reportable}
}
\description{
Replace a report in \code{teal_reportable}
}
