% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_report-report.R
\name{report}
\alias{report}
\title{Extract report from \code{teal_report}}
\usage{
report(x)
}
\arguments{
\item{x}{(\code{teal_report})}
}
\value{
\code{teal_report}
}
\description{
Extract report from \code{teal_report}
}
