% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadReporterModule.R
\name{report_load_ui}
\alias{report_load_ui}
\title{User Interface to Load \code{Reporter}}
\usage{
report_load_ui(id, label = NULL)
}
\arguments{
\item{id}{\code{character(1)} this \code{shiny} module's id.}

\item{label}{(\code{character(1)}) label before the icon. By default \code{NULL}.}
}
\value{
\code{shiny::tagList}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Button to upload \code{ReporterCard}(s) to the \code{Reporter}.

For more details see the vignette: \code{vignette("simpleReporter", "teal.reporter")}.
}
