% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadModule.R
\name{report_render_and_compress}
\alias{report_render_and_compress}
\title{Render the report}
\usage{
report_render_and_compress(
  reporter,
  rmd_yaml_args,
  global_knitr,
  file = tempdir()
)
}
\arguments{
\item{reporter}{(\code{Reporter}) instance.}

\item{rmd_yaml_args}{(\verb{named list}) with \code{Rmd} \code{yaml} header fields and their values.}

\item{global_knitr}{(\code{list}) a global \code{knitr} parameters, like echo.
But if local parameter is set it will have priority.}

\item{file}{(\code{character(1)}) where to copy the returned directory.}
}
\value{
\code{file} argument, invisibly.
}
\description{
Render the report and zip the created directory.
}
\keyword{internal}
