% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportCard.R
\name{ReportCard}
\alias{ReportCard}
\title{\code{ReportCard}}
\description{
\code{ReportCard}

\code{ReportCard}
}
\examples{

## ------------------------------------------------
## Method `ReportCard$new`
## ------------------------------------------------

card <- ReportCard$new()


## ------------------------------------------------
## Method `ReportCard$append_table`
## ------------------------------------------------

card <- ReportCard$new()$append_table(iris)


## ------------------------------------------------
## Method `ReportCard$append_plot`
## ------------------------------------------------

card <- ReportCard$new()$append_plot(
  ggplot2::ggplot(iris, ggplot2::aes(x = Petal.Length)) + ggplot2::geom_histogram()
)


## ------------------------------------------------
## Method `ReportCard$append_text`
## ------------------------------------------------

card <- ReportCard$new()$append_text("A paragraph of default text")


## ------------------------------------------------
## Method `ReportCard$get_content`
## ------------------------------------------------

card <- ReportCard$new()$append_text("Some text")$append_plot(
  ggplot2::ggplot(iris, ggplot2::aes(x = Petal.Length)) + ggplot2::geom_histogram()
)
card$get_content()

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ReportCard$new()}}
\item \href{#method-append_table}{\code{ReportCard$append_table()}}
\item \href{#method-append_plot}{\code{ReportCard$append_plot()}}
\item \href{#method-append_text}{\code{ReportCard$append_text()}}
\item \href{#method-get_content}{\code{ReportCard$get_content()}}
\item \href{#method-deep_clone}{\code{ReportCard$deep_clone()}}
\item \href{#method-clone}{\code{ReportCard$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Returns a \code{ReportCard} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportCard$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a \code{ReportCard} object
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{card <- ReportCard$new()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-append_table"></a>}}
\if{latex}{\out{\hypertarget{method-append_table}{}}}
\subsection{Method \code{append_table()}}{
Appends a table to this \code{ReportCard}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportCard$append_table(table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{table}}{the appended table}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisibly self
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{card <- ReportCard$new()$append_table(iris)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-append_plot"></a>}}
\if{latex}{\out{\hypertarget{method-append_plot}{}}}
\subsection{Method \code{append_plot()}}{
Appends a plot to this \code{ReportCard}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportCard$append_plot(plot)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plot}}{the appended plot}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisibly self
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{card <- ReportCard$new()$append_plot(
  ggplot2::ggplot(iris, ggplot2::aes(x = Petal.Length)) + ggplot2::geom_histogram()
)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-append_text"></a>}}
\if{latex}{\out{\hypertarget{method-append_text}{}}}
\subsection{Method \code{append_text()}}{
Appends a paragraph of text to this \code{ReportCard}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportCard$append_text(text, style = TextBlock$new()$get_available_styles()[1])}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{(\code{character(0)} or \code{character(1)}) the text}

\item{\code{style}}{(\code{character(1)}) the style of the paragraph. One of: \code{default}, \code{header}, \code{verbatim}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisibly self
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{card <- ReportCard$new()$append_text("A paragraph of default text")

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_content"></a>}}
\if{latex}{\out{\hypertarget{method-get_content}{}}}
\subsection{Method \code{get_content()}}{
Returns the content of this \code{ReportCard}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportCard$get_content()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list()} list of \code{TableBlock}, \code{TextBlock} and \code{PictureBlock}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{card <- ReportCard$new()$append_text("Some text")$append_plot(
  ggplot2::ggplot(iris, ggplot2::aes(x = Petal.Length)) + ggplot2::geom_histogram()
)
card$get_content()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deep_clone"></a>}}
\if{latex}{\out{\hypertarget{method-deep_clone}{}}}
\subsection{Method \code{deep_clone()}}{
The copy constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportCard$deep_clone(name, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{the name of the field}

\item{\code{value}}{the value of the field}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the new value of the field
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportCard$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
