% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Previewer.R
\name{reporter_previewer}
\alias{reporter_previewer}
\alias{preview_report_button_ui}
\alias{preview_report_button_srv}
\title{Show report previewer button module}
\usage{
preview_report_button_ui(id, label = "Preview Report")

preview_report_button_srv(id, reporter)
}
\arguments{
\item{id}{(\code{character(1)}) \code{shiny} module instance id.}

\item{label}{(\code{character(1)}) label of the button. By default it is "Preview Report".}

\item{reporter}{(\code{Reporter}) instance.}
}
\value{
\code{NULL}.
}
\description{
Provides a button that triggers showing the report preview in a modal.
}
\details{
For more details see the vignette: \code{vignette("previewerReporter", "teal.reporter")}.
}
