% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Previewer.R
\name{reporter_previewer_srv}
\alias{reporter_previewer_srv}
\title{Reporter Previewer Server}
\usage{
reporter_previewer_srv(
  id,
  reporter,
  rmd_yaml_args = list(author = "NEST", title = "Report", date =
    as.character(Sys.Date()), output = "html_document")
)
}
\arguments{
\item{id}{\code{character(1)} this \code{shiny} module's id.}

\item{reporter}{\code{Reporter} instance}

\item{rmd_yaml_args}{\verb{named list} vector with \code{Rmd} \code{yaml} header fields and their default values.
Default \code{list(author = "NEST", title = "Report", date = Sys.Date(), output = "html_document")}.
Please update only values at this moment.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
server supporting the functionalities of the reporter previewer
For more details see the vignette: \code{vignette("previewerReporter", "teal.reporter")}.
}
