% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Previewer.R
\name{reporter_previewer_ui}
\alias{reporter_previewer_ui}
\title{Reporter Previewer User Interface}
\usage{
reporter_previewer_ui(
  id,
  rmd_output = c("html_document", "pdf_document", "powerpoint_presentation",
    "word_document"),
  rmd_yaml_args = list(author = "NEST", title = "Report", date =
    as.character(Sys.Date()), output = "html_document")
)
}
\arguments{
\item{id}{\code{character(1)} \code{shiny} module id.}

\item{rmd_output}{\code{character} vector with \code{rmarkdown} output types,
by default all possible \code{c("pdf_document", "html_document", "powerpoint_presentation", "word_document")}.}

\item{rmd_yaml_args}{\verb{named list} vector with \code{Rmd} \code{yaml} header fields and their default values.
Default \code{list(author = "NEST", title = "Report", date = Sys.Date(), output = "html_document")}.
Please update only values at this moment.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
reporter previewer user interface to visualize and manipulate the already added report Cards
}
