% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResetModule.R
\name{reset_report_button_ui}
\alias{reset_report_button_ui}
\title{Reset Button Reporter User Interface}
\usage{
reset_report_button_ui(id, label = NULL)
}
\arguments{
\item{id}{\code{character(1)} this \code{shiny} module's id.}

\item{label}{\code{character(1)} label before icon, if used then dynamic hover label is not available.
By default \code{NULL} so a dynamic hover label is used.}
}
\value{
\code{shiny::tagList}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
button for resetting the report content.

For more details see the vignette: \code{vignette("simpleReporter", "teal.reporter")}.
}
