% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_utils.R
\name{rmd_output_arguments}
\alias{rmd_output_arguments}
\title{Get document output arguments from the \code{rmarkdown} package}
\usage{
rmd_output_arguments(output_name, default_values = FALSE)
}
\arguments{
\item{output_name}{(\code{character}) \code{rmarkdown} output name.}

\item{default_values}{(\code{logical(1)}) if to return a default values for each argument.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Retrieves the arguments for a specified document output type from the \code{rmarkdown} package.
}
\examples{
rmd_output_arguments("pdf_document")
rmd_output_arguments("pdf_document", TRUE)
}
