% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_utils.R
\name{rmd_outputs}
\alias{rmd_outputs}
\title{Get document output types from the rmarkdown package}
\usage{
rmd_outputs()
}
\value{
\code{character} vector.
}
\description{
get document output types from the \code{rmarkdown} package.
}
\examples{
rmd_outputs()

}
