% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleReporter.R
\name{simple_reporter_srv}
\alias{simple_reporter_srv}
\title{Simple Reporter Server}
\usage{
simple_reporter_srv(id, reporter, card_fun, notification = TRUE)
}
\arguments{
\item{id}{\code{character}}

\item{reporter}{\code{Reporter} instance.}

\item{card_fun}{\code{function} which returns a \code{ReportCard} instance,
the function have at least two arguments \code{card} and \code{comment}.}

\item{notification}{logical if to add shiny notification about the download process.}
}
\value{
\code{shiny::moduleServer}
}
\description{
two buttons for adding views and downloading the Report.
The add module has \code{addReportCard} id and download module the \code{downloadButton} id.

For more details see the vignette: \code{vignette("simpleReporter", "teal.reporter")}.
}
