% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleReporter.R
\name{simple_reporter_srv}
\alias{simple_reporter_srv}
\title{Simple Reporter Server}
\usage{
simple_reporter_srv(id, reporter, card, notification = TRUE)
}
\arguments{
\item{id}{\code{character}}

\item{reporter}{\code{Reporter} instance.}

\item{card}{\code{ReportCard} instance}

\item{notification}{logical if to add shiny notification about the download process.}
}
\value{
\code{shiny::tagList}
}
\description{
two buttons for adding views and downloading the Report.
The add module has \code{addReportCard} id and download module the \code{downloadButton} id.

For more details see the vignette: \code{vignette("simpleReporter", "teal.reporter")}.
}
