% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleReporter.R
\name{simple_reporter_srv}
\alias{simple_reporter_srv}
\title{Simple Reporter Server}
\usage{
simple_reporter_srv(id, reporter, card_fun)
}
\arguments{
\item{id}{\code{character(1)} this \code{shiny} module's id.}

\item{reporter}{\code{\link{Reporter}} instance.}

\item{card_fun}{\code{function} which returns a \code{\link{ReportCard}} instance,
the function have at\code{card}argument and optional \code{comment}.}
}
\value{
\code{shiny::moduleServer}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
three buttons for adding cards, downloading and resetting the Report.
The add module has \code{add_report_card_simple} id, the download module the \code{download_button_simple} id
and the reset module the \code{reset_button_simple} id.

For more details see the vignette: \code{vignette("simpleReporter", "teal.reporter")}.
}
