% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleReporter.R
\name{simple_reporter_ui}
\alias{simple_reporter_ui}
\title{Simple Reporter User Interface}
\usage{
simple_reporter_ui(id)
}
\arguments{
\item{id}{\code{character}}
}
\value{
\code{shiny::tagList}
}
\description{
three buttons for adding views, downloading and resetting the Report.

For more details see the vignette: \code{vignette("simpleReporter", "teal.reporter")}.
}
\examples{
if (interactive()) {
  shiny::shinyApp(
    ui = function(id) {
      simple_reporter_ui("simple")
    },
    server = function(input, output, session) {
      simple_reporter_srv("simple", Reporter$new(), function(card) card)
    }
  )
}
}
