% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleReporter.R
\name{simple_reporter_ui}
\alias{simple_reporter_ui}
\title{Simple Reporter User Interface}
\usage{
simple_reporter_ui(id, label = "")
}
\arguments{
\item{id}{\code{character(1)} this \code{shiny} module's id.}

\item{label}{\code{character(1)} label before the icon.}
}
\value{
\code{shiny.tag}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
three buttons for adding cards, downloading and resetting the Report.

For more details see the vignette: \code{vignette("simpleReporter", "teal.reporter")}.
}
\examples{
if (interactive()) {
  shiny::shinyApp(
    ui = shiny::fluidPage(simple_reporter_ui("simple")),
    server = function(input, output, session) {
      simple_reporter_srv("simple", Reporter$new(), function(card) card)
    }
  )
}
}
