% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleReporter.R
\name{simple_reporter_ui}
\alias{simple_reporter_ui}
\title{Simple Reporter User Interface}
\usage{
simple_reporter_ui(id)
}
\arguments{
\item{id}{character}
}
\value{
\code{shiny::tagList}
}
\description{
two buttons for adding views and downloading the Report.

For more details see the vignette: \code{vignette("simpleReporter", "teal.reporter")}.
}
