% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Editor.R
\name{ui_editor_block}
\alias{ui_editor_block}
\alias{srv_editor_block}
\title{UI and Server functions for editing report document blocks}
\usage{
ui_editor_block(id, value)

srv_editor_block(id, value)
}
\arguments{
\item{id}{(\code{character(1)}) A unique identifier for the module.}

\item{value}{The content of the block to be edited. It can be a character string or other types.}
}
\description{
These functions provide a user interface and server logic for editing and extending
the editor functionality to support new data types.
}
