% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableBlock.R
\name{TableBlock}
\alias{TableBlock}
\title{\code{TableBlock}}
\description{
\code{TableBlock}

\code{TableBlock}
}
\examples{

## ------------------------------------------------
## Method `TableBlock$set_content`
## ------------------------------------------------

block <- teal.reporter:::TableBlock$new()
block$set_content(iris)

}
\keyword{internal}
\section{Super class}{
\code{\link[teal.reporter:ContentBlock]{teal.reporter::ContentBlock}} -> \code{TableBlock}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TableBlock-new}{\code{TableBlock$new()}}
\item \href{#method-TableBlock-set_content}{\code{TableBlock$set_content()}}
\item \href{#method-TableBlock-finalize}{\code{TableBlock$finalize()}}
\item \href{#method-TableBlock-from_list}{\code{TableBlock$from_list()}}
\item \href{#method-TableBlock-to_list}{\code{TableBlock$to_list()}}
\item \href{#method-TableBlock-clone}{\code{TableBlock$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="teal.reporter" data-topic="ContentBlock" data-id="get_content"><a href='../../teal.reporter/html/ContentBlock.html#method-ContentBlock-get_content'><code>teal.reporter::ContentBlock$get_content()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableBlock-new"></a>}}
\if{latex}{\out{\hypertarget{method-TableBlock-new}{}}}
\subsection{Method \code{new()}}{
Returns a new \code{TableBlock} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableBlock$new(table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{table}}{(\code{data.frame}, \code{DT}, \code{rtables}, \code{TableTree}) a table assigned to this \code{TableBlock}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a \code{TableBlock} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableBlock-set_content"></a>}}
\if{latex}{\out{\hypertarget{method-TableBlock-set_content}{}}}
\subsection{Method \code{set_content()}}{
Sets content of this \code{TableBlock}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableBlock$set_content(content)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{content}}{(\code{data.frame}, \code{rtables}) a table assigned to this \code{TableBlock}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
throws if argument is not a table-like object.
}

\subsection{Returns}{
invisibly self
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{block <- teal.reporter:::TableBlock$new()
block$set_content(iris)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableBlock-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-TableBlock-finalize}{}}}
\subsection{Method \code{finalize()}}{
finalize of this \code{TableBlock}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableBlock$finalize()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Removes the temporary file created in the constructor.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableBlock-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-TableBlock-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableBlock$from_list(x, base_path = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableBlock-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-TableBlock-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableBlock$to_list(base_path = ".")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableBlock-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TableBlock-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableBlock$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
