% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal.reporter-package.R
\docType{package}
\name{teal.reporter-package}
\alias{teal.reporter}
\alias{teal.reporter-package}
\title{Reporting tools for \code{shiny} modules.}
\description{
A collection of tools for creating, previewing and modifying reports from \code{shiny} modules.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/insightsengineering/teal.reporter}
  \item \url{https://insightsengineering.github.io/teal.reporter/}
  \item Report bugs at \url{https://github.com/insightsengineering/teal.reporter/issues}
}

}
\author{
\strong{Maintainer}: Maciej Nasinski \email{maciej.nasinski@contractors.roche.com}

Authors:
\itemize{
  \item Konrad Pagacz \email{konrad.pagacz@contractors.roche.com}
  \item Mahmoud Hallal \email{mahmoud.hallal@roche.com}
}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
\keyword{internal}
