% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal.reporter.R
\docType{package}
\name{teal.reporter-package}
\alias{teal.reporter}
\alias{teal.reporter-package}
\title{Reporting tools for \code{shiny} modules.}
\description{
A collection of tools for creating, previewing and modifying reports from \code{shiny} modules.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/insightsengineering/teal.reporter}
  \item \url{https://insightsengineering.github.io/teal.reporter/}
  \item Report bugs at \url{https://github.com/insightsengineering/teal.reporter/issues}
}

}
\author{
\strong{Maintainer}: Dawid Kaledkowski \email{dawid.kaledkowski@roche.com} (\href{https://orcid.org/0000-0001-9533-457X}{ORCID})

Authors:
\itemize{
  \item Aleksander Chlebowski \email{aleksander.chlebowski@contractors.roche.com} (\href{https://orcid.org/0000-0001-5018-6294}{ORCID})
  \item Andre Verissimo \email{andre.verissimo@roche.com} (\href{https://orcid.org/0000-0002-2212-339X}{ORCID})
  \item Kartikeya Kirar \email{kartikeya.kirar@businesspartner.roche.com} (\href{https://orcid.org/0009-0005-1258-4757}{ORCID})
  \item Vedha Viyash \email{vedha.viyash@roche.com}
  \item Marcin Kosinski \email{marcin.kosinski.mk1@roche.com}
  \item Maciej Nasinski
  \item Konrad Pagacz
  \item Mahmoud Hallal \email{mahmoud.hallal@roche.com}
}

Other contributors:
\itemize{
  \item Chendi Liao \email{chendi.liao@roche.com} [reviewer]
  \item Dony Unardi \email{unardid@gene.com} [reviewer]
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
\keyword{internal}
