% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal.reporter.R
\docType{package}
\name{teal.reporter}
\alias{teal.reporter-package}
\alias{teal.reporter}
\title{Reporting tools for \code{shiny} modules.}
\description{
A collection of tools for creating, previewing and modifying reports from \code{shiny} modules.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/insightsengineering/teal.reporter}
  \item \url{https://insightsengineering.github.io/teal.reporter/}
  \item Report bugs at \url{https://github.com/insightsengineering/teal.reporter/issues}
}

}
\author{
\strong{Maintainer}: Dawid Kaledkowski \email{dawid.kaledkowski@roche.com}

Authors:
\itemize{
  \item Maciej Nasinski
  \item Konrad Pagacz
  \item Mahmoud mahmoud.hallal@roche.com Hallal
}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
\keyword{internal}
