% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal.reporter.R
\docType{package}
\name{teal.reporter}
\alias{teal.reporter}
\title{Reporting tools for \code{shiny} modules.}
\description{
A collection of tools for creating, previewing and modifying reports from \code{shiny} modules.
}
