% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_report-class.R
\docType{class}
\name{teal_report-class}
\alias{teal_report-class}
\title{Reproducible report}
\description{
Reproducible report container class. Inherits code tracking behavior from \code{\link[teal.data:teal_data-class]{teal.data::teal_data}}.
}
\details{
This class provides an isolated environment in which to store and process data with all code being recorded.
The environment, code, data set names, and data joining keys are stored in their respective slots.
These slots should never be accessed directly, use the provided get/set functions.

As code is evaluated in \code{teal_data}, messages and warnings are stored in their respective slots.
If errors are raised, a \code{qenv.error} object is returned.
}
\section{Slots}{

\describe{
\item{\code{.xData}}{(\code{environment}) environment containing data sets and possibly
auxiliary variables.
Access variables with \code{\link[=get]{get()}}, \code{\link{$}}  or [\code{[[}].
No setter provided. Evaluate code to add variables into \verb{@.xData}.}

\item{\code{code}}{(\code{list} of \code{character}) representing code necessary to reproduce the contents of \code{qenv}.
Access with \code{\link[teal.code:get_code]{teal.code::get_code()}}.
No setter provided. Evaluate code to append code to the slot.}

\item{\code{join_keys}}{(\code{join_keys}) object specifying joining keys for data sets in
\verb{@.xData}.
Access or modify with \code{\link[teal.data:join_keys]{teal.data::join_keys()}}.}

\item{\code{verified}}{(\code{logical(1)}) flag signifying that code in \verb{@code} has been
proven to yield contents of \verb{@.xData}.
Used internally. See \code{\link[teal.data:verify]{teal.data::verify()}} for more details.}

\item{\code{card}}{(\code{teal_card})}
}}

\keyword{internal}
