% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TealReportCard.R
\name{TealReportCard}
\alias{TealReportCard}
\title{\code{TealReportCard}}
\description{
A child of \code{ReportCard} that is used for teal specific applications.
In addition to the parent methods, it supports rendering teal specific elements such as
the source code, the encodings panel content and the filter panel content as part of the
meta data.
}
\examples{

## ------------------------------------------------
## Method `TealReportCard$append_src`
## ------------------------------------------------

card <- TealReportCard$new()$append_src(
  "ggplot2::ggplot(iris, ggplot2::aes(x = Petal.Length)) + ggplot2::geom_histogram()"
)


## ------------------------------------------------
## Method `TealReportCard$append_fs`
## ------------------------------------------------

card <- TealReportCard$new()$append_fs(
  list(data = list(X = list(selected = c(1, 10))))
)


## ------------------------------------------------
## Method `TealReportCard$append_encodings`
## ------------------------------------------------

card <- TealReportCard$new()$append_encodings(list("variable 1 is X"))

}
\section{Super class}{
\code{\link[teal.reporter:ReportCard]{teal.reporter::ReportCard}} -> \code{TealReportCard}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-append_src}{\code{TealReportCard$append_src()}}
\item \href{#method-append_fs}{\code{TealReportCard$append_fs()}}
\item \href{#method-append_encodings}{\code{TealReportCard$append_encodings()}}
\item \href{#method-clone}{\code{TealReportCard$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="teal.reporter" data-topic="ReportCard" data-id="append_metadata">}\href{../../teal.reporter/html/ReportCard.html#method-append_metadata}{\code{teal.reporter::ReportCard$append_metadata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.reporter" data-topic="ReportCard" data-id="append_plot">}\href{../../teal.reporter/html/ReportCard.html#method-append_plot}{\code{teal.reporter::ReportCard$append_plot()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.reporter" data-topic="ReportCard" data-id="append_table">}\href{../../teal.reporter/html/ReportCard.html#method-append_table}{\code{teal.reporter::ReportCard$append_table()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.reporter" data-topic="ReportCard" data-id="append_text">}\href{../../teal.reporter/html/ReportCard.html#method-append_text}{\code{teal.reporter::ReportCard$append_text()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.reporter" data-topic="ReportCard" data-id="get_content">}\href{../../teal.reporter/html/ReportCard.html#method-get_content}{\code{teal.reporter::ReportCard$get_content()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.reporter" data-topic="ReportCard" data-id="get_deparsers">}\href{../../teal.reporter/html/ReportCard.html#method-get_deparsers}{\code{teal.reporter::ReportCard$get_deparsers()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.reporter" data-topic="ReportCard" data-id="initialize">}\href{../../teal.reporter/html/ReportCard.html#method-initialize}{\code{teal.reporter::ReportCard$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-append_src"></a>}}
\if{latex}{\out{\hypertarget{method-append_src}{}}}
\subsection{Method \code{append_src()}}{
Appends the source code to the \code{metadata} of this \code{TealReportCard}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealReportCard$append_src(
  src,
  deparse = function(x) paste0("```\\n", x, "\\n```\\n")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{src}}{(\code{character(1)}) code as text}

\item{\code{deparse}}{(\code{function}) to convert a src to a string, by default \code{function(x) paste0("```\\n", x, "\\n```\\n")}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisibly self
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{card <- TealReportCard$new()$append_src(
  "ggplot2::ggplot(iris, ggplot2::aes(x = Petal.Length)) + ggplot2::geom_histogram()"
)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-append_fs"></a>}}
\if{latex}{\out{\hypertarget{method-append_fs}{}}}
\subsection{Method \code{append_fs()}}{
Appends the filter state list to the \code{metadata} of this \code{TealReportCard}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealReportCard$append_fs(fs, deparse = deparse1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fs}}{(\code{list}) list of filter states.}

\item{\code{deparse}}{(\code{function}) to convert a fs to a string, by default \code{base::deparse1}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisibly self
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{card <- TealReportCard$new()$append_fs(
  list(data = list(X = list(selected = c(1, 10))))
)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-append_encodings"></a>}}
\if{latex}{\out{\hypertarget{method-append_encodings}{}}}
\subsection{Method \code{append_encodings()}}{
Appends the encodings list to the \code{metadata} of this \code{TealReportCard}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealReportCard$append_encodings(encodings, deparse = deparse1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{encodings}}{(\code{list}) list of encodings selections of the teal app}

\item{\code{deparse}}{(\code{function}) to convert a encodings to a string, by default \code{base::deparse1}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisibly self
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{card <- TealReportCard$new()$append_encodings(list("variable 1 is X"))

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealReportCard$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
