% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextBlock.R
\name{TextBlock}
\alias{TextBlock}
\title{\code{TextBlock}}
\description{
\code{TextBlock}

\code{TextBlock}
}
\examples{

## ------------------------------------------------
## Method `TextBlock$new`
## ------------------------------------------------

block <- teal.reporter:::TextBlock$new()


## ------------------------------------------------
## Method `TextBlock$set_style`
## ------------------------------------------------

block <- teal.reporter:::TextBlock$new()
block$set_style("header2")


## ------------------------------------------------
## Method `TextBlock$get_style`
## ------------------------------------------------

block <- teal.reporter:::TextBlock$new()
block$get_style()


## ------------------------------------------------
## Method `TextBlock$get_available_styles`
## ------------------------------------------------

block <- teal.reporter:::TextBlock$new()
block$get_available_styles()

}
\keyword{internal}
\section{Super class}{
\code{\link[teal.reporter:ContentBlock]{teal.reporter::ContentBlock}} -> \code{TextBlock}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TextBlock$new()}}
\item \href{#method-set_style}{\code{TextBlock$set_style()}}
\item \href{#method-get_style}{\code{TextBlock$get_style()}}
\item \href{#method-get_available_styles}{\code{TextBlock$get_available_styles()}}
\item \href{#method-clone}{\code{TextBlock$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="teal.reporter" data-topic="ContentBlock" data-id="deep_clone">}\href{../../teal.reporter/html/ContentBlock.html#method-deep_clone}{\code{teal.reporter::ContentBlock$deep_clone()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.reporter" data-topic="ContentBlock" data-id="get_content">}\href{../../teal.reporter/html/ContentBlock.html#method-get_content}{\code{teal.reporter::ContentBlock$get_content()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.reporter" data-topic="ContentBlock" data-id="set_content">}\href{../../teal.reporter/html/ContentBlock.html#method-set_content}{\code{teal.reporter::ContentBlock$set_content()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Returns a \code{TextBlock} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextBlock$new(content = character(0), style = private$styles[1])}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{content}}{(\code{character(1)} or \code{character(0)}) a string assigned to this \code{TextBlock}}

\item{\code{style}}{(\code{character(1)}) one of: \code{default}, \code{header}, \code{verbatim}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Returns a \code{TextBlock} object with no content and the default style.
}

\subsection{Returns}{
\code{TextBlock}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{block <- teal.reporter:::TextBlock$new()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_style"></a>}}
\if{latex}{\out{\hypertarget{method-set_style}{}}}
\subsection{Method \code{set_style()}}{
Sets the style of this \code{TextBlock}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextBlock$set_style(style)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{style}}{(\code{character(1)}) one of: \code{default}, \code{header2}, \code{header3}, \code{verbatim}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The style has bearing on the rendering of this block.
}

\subsection{Returns}{
invisibly self
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{block <- teal.reporter:::TextBlock$new()
block$set_style("header2")

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_style"></a>}}
\if{latex}{\out{\hypertarget{method-get_style}{}}}
\subsection{Method \code{get_style()}}{
Returns the style of this \code{TextBlock}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextBlock$get_style()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character(1)} the style of this \code{TextBlock}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{block <- teal.reporter:::TextBlock$new()
block$get_style()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_available_styles"></a>}}
\if{latex}{\out{\hypertarget{method-get_available_styles}{}}}
\subsection{Method \code{get_available_styles()}}{
Returns an array of styles available to this \code{TextBlock}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextBlock$get_available_styles()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a \code{character} array of styles
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{block <- teal.reporter:::TextBlock$new()
block$get_available_styles()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextBlock$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
