% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextBlock.R
\name{TextBlock}
\alias{TextBlock}
\title{\code{TextBlock}}
\description{
\code{TextBlock}

\code{TextBlock}
}
\examples{

## ------------------------------------------------
## Method `TextBlock$new`
## ------------------------------------------------

block <- teal.reporter:::TextBlock$new()


## ------------------------------------------------
## Method `TextBlock$set_style`
## ------------------------------------------------

block <- teal.reporter:::TextBlock$new()
block$set_style("header2")


## ------------------------------------------------
## Method `TextBlock$get_style`
## ------------------------------------------------

block <- teal.reporter:::TextBlock$new()
block$get_style()


## ------------------------------------------------
## Method `TextBlock$get_available_styles`
## ------------------------------------------------

block <- teal.reporter:::TextBlock$new()
block$get_available_styles()


## ------------------------------------------------
## Method `TextBlock$from_list`
## ------------------------------------------------

block <- teal.reporter:::TextBlock$new()
block$from_list(list(text = "sth", style = "default"))

## ------------------------------------------------
## Method `TextBlock$to_list`
## ------------------------------------------------

block <- teal.reporter:::TextBlock$new()
block$to_list()
}
\keyword{internal}
\section{Super class}{
\code{\link[teal.reporter:ContentBlock]{teal.reporter::ContentBlock}} -> \code{TextBlock}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TextBlock-new}{\code{TextBlock$new()}}
\item \href{#method-TextBlock-set_style}{\code{TextBlock$set_style()}}
\item \href{#method-TextBlock-get_style}{\code{TextBlock$get_style()}}
\item \href{#method-TextBlock-get_available_styles}{\code{TextBlock$get_available_styles()}}
\item \href{#method-TextBlock-from_list}{\code{TextBlock$from_list()}}
\item \href{#method-TextBlock-to_list}{\code{TextBlock$to_list()}}
\item \href{#method-TextBlock-clone}{\code{TextBlock$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="teal.reporter" data-topic="ContentBlock" data-id="get_content"><a href='../../teal.reporter/html/ContentBlock.html#method-ContentBlock-get_content'><code>teal.reporter::ContentBlock$get_content()</code></a></li>
<li><span class="pkg-link" data-pkg="teal.reporter" data-topic="ContentBlock" data-id="set_content"><a href='../../teal.reporter/html/ContentBlock.html#method-ContentBlock-set_content'><code>teal.reporter::ContentBlock$set_content()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TextBlock-new"></a>}}
\if{latex}{\out{\hypertarget{method-TextBlock-new}{}}}
\subsection{Method \code{new()}}{
Returns a \code{TextBlock} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextBlock$new(content = character(0), style = private$styles[1])}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{content}}{(\code{character(1)} or \code{character(0)}) a string assigned to this \code{TextBlock}}

\item{\code{style}}{(\code{character(1)}) one of: \code{"default"}, \code{"header2"}, \code{"header3"} \code{"verbatim"}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Returns a \code{TextBlock} object with no content and the default style.
}

\subsection{Returns}{
\code{TextBlock}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{block <- teal.reporter:::TextBlock$new()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TextBlock-set_style"></a>}}
\if{latex}{\out{\hypertarget{method-TextBlock-set_style}{}}}
\subsection{Method \code{set_style()}}{
Sets the style of this \code{TextBlock}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextBlock$set_style(style)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{style}}{(\code{character(1)}) one of: \code{"default"}, \code{"header2"}, \code{"header3"} \code{"verbatim"}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The style has bearing on the rendering of this block.
}

\subsection{Returns}{
invisibly self
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{block <- teal.reporter:::TextBlock$new()
block$set_style("header2")

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TextBlock-get_style"></a>}}
\if{latex}{\out{\hypertarget{method-TextBlock-get_style}{}}}
\subsection{Method \code{get_style()}}{
Returns the style of this \code{TextBlock}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextBlock$get_style()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character(1)} the style of this \code{TextBlock}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{block <- teal.reporter:::TextBlock$new()
block$get_style()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TextBlock-get_available_styles"></a>}}
\if{latex}{\out{\hypertarget{method-TextBlock-get_available_styles}{}}}
\subsection{Method \code{get_available_styles()}}{
Returns an array of styles available to this \code{TextBlock}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextBlock$get_available_styles()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a \code{character} array of styles
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{block <- teal.reporter:::TextBlock$new()
block$get_available_styles()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TextBlock-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-TextBlock-from_list}{}}}
\subsection{Method \code{from_list()}}{
Create the \code{TextBlock} from a list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextBlock$from_list(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\verb{named list} with two fields \code{c("text", "style")}.
Use the \code{get_available_styles} method to get all possible styles.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisibly self
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{block <- teal.reporter:::TextBlock$new()
block$from_list(list(text = "sth", style = "default"))
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TextBlock-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-TextBlock-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert the \code{TextBlock} to a list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextBlock$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\verb{named list} with a text and style.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{block <- teal.reporter:::TextBlock$new()
block$to_list()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TextBlock-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TextBlock-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextBlock$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
