% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_flextable}
\alias{to_flextable}
\title{Convert to Flextable}
\usage{
to_flextable(content)
}
\arguments{
\item{content}{Supported formats: "data.frame", "rtables", "TableTree", "ElementaryTable"}
}
\description{
Convert content into a flextable, merge cells with colspan > 1,
align columns to the center, and row names to the left.
Indent the row names by 10 times indentation.
}
