% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_flextable}
\alias{to_flextable}
\title{Convert content into a \code{flextable}}
\usage{
to_flextable(content)
}
\arguments{
\item{content}{Supported formats: \code{data.frame}, \code{rtables}, \code{TableTree}, \code{ElementaryTable}, \code{listing_df}}
}
\value{
\code{flextable}.
}
\description{
Function transform content into a \code{flextable}, merge cells with \code{colspan} > 1
align columns to the center, and row names to the left
indent the row names by 10 times indentation.
}
\keyword{internal}
