% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadModule.R
\name{to_rmd}
\alias{to_rmd}
\title{Convert \code{Reporter}/\code{card} content to \code{rmarkdown}}
\usage{
to_rmd(block, output_dir, ...)
}
\arguments{
\item{block}{(\code{any}) content which can be represented in \code{rmarkdown} syntax}

\item{output_dir}{(\code{character(1)}) path to the directory where files should be written to.}
}
\value{
\code{character(1)} containing a content or \code{rmarkdown} document
}
\description{
Convert \code{Reporter}/\code{card} content to \code{rmarkdown}
}
\keyword{internal}
