% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_rmd.R
\name{to_rmd}
\alias{to_rmd}
\title{Convert \code{ReporterCard}/\code{teal_card} content to \code{rmarkdown}}
\usage{
to_rmd(block, output_dir, ...)
}
\arguments{
\item{block}{(\code{any}) content which can be represented in Rmarkdown syntax.}

\item{output_dir}{(\code{character(1)}) path to the directory where files should be written to. Beware
that absolute paths will break a reproducibility of the Rmarkdown document.}
}
\value{
\code{character(1)} containing a content or Rmarkdown document.
}
\description{
This is an S3 generic that is used to generate content in \code{rmarkdown} format
from various types of blocks in a \code{ReporterCard} or \code{teal_card} object.
}
\section{Customise \code{to_rmd}}{
The methods for this S3 generic can be extended by the app developer or even overwritten.
For this a function with the name \verb{to_rmd.<class>} should be defined in the
Global Environment, where \verb{<class>} is the class of the object to be converted.

For example, to override the default behavior for \code{code_chunk} class, you can use:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{to_rmd.code_chunk <- function(block, output_dir, ..., output_format) \{
  # custom implementation
  sprintf("### A custom code chunk\\n\\n```\{r\}\\n\%s\\n```\\n", block)
\}
}\if{html}{\out{</div>}}

Alternatively, you can register the S3 method using \code{registerS3method("to_rmd", "<class>", fun)}
}

\keyword{internal}
