% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_utils.R
\name{yaml_quoted}
\alias{yaml_quoted}
\title{quoted string for \code{yaml}}
\usage{
yaml_quoted(x)
}
\arguments{
\item{x}{\code{character}}
}
\description{
add quoted attribute for \code{yaml} package
}
\keyword{internal}
