% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_utils.R
\name{yaml_quoted}
\alias{yaml_quoted}
\title{quoted string for \code{yaml}}
\usage{
yaml_quoted(x)
}
\arguments{
\item{x}{\code{character}}
}
\description{
add quoted attribute for \code{yaml} package
}
\examples{
yaml <- list(
  author = teal.reporter:::yaml_quoted("NEST"),
  title = teal.reporter:::yaml_quoted("Report"),
  date = teal.reporter:::yaml_quoted("07/04/2019"),
  output = list(pdf_document = list(keep_tex = TRUE))
)
yaml::as.yaml(yaml)
}
\keyword{internal}
